% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNewSecret.R
\name{addNewSecret}
\alias{addNewSecret}
\title{Online addition of key-value pairs.}
\usage{
addNewSecret(name, users, vault)
}
\value{
The value stored.
}
\description{
Online addition of key-value pairs.
}
\examples{

}
