% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNewUserToSecretVault.R
\name{addNewUserToSecretVault}
\alias{addNewUserToSecretVault}
\title{Add users who may access the vault.}
\usage{
addNewUserToSecretVault(username)
}
\value{
A
}
\description{
Add users who may access the vault.
}
\examples{
  addNewUserToSecretVaule( 'alice' )
}
