% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{AddNote}
\alias{AddNote}
\title{Add a note to the counts note dataframe}
\usage{
AddNote(note, reference, countsObject)
}
\arguments{
\item{note}{a string containing the note to be added.}

\item{reference}{the reference of the sequence to take a note about.}

\item{countsObject}{dataframe with columns 1:5 with sequence info (unique sequence references, notes, Swedish name, Latin name and percentage out of all sequences in the data frame) and the rest of the columns of samples with counts of the sequences.}
}
\value{
dataframe with columns 1:5 with sequence info (sequence reference, notes, Swedish name, Latin name and percentage out of all sequences in the data frame) and the rest of the columns of samples with counts of the sequences.
}
\description{
Add a note to the counts note dataframe
}
