% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{all_names}
\alias{all_names}
\title{A list of species found in Sweden}
\format{
\subsection{\code{all_names}}{

A data frame with 24528 rows and 2 columns:
\describe{
\item{Latin}{Scientific name of species}
\item{Swedish}{Swedish name of species}
}
}
}
\source{
\url{https://namnochslaktskap.artfakta.se/}
}
\usage{
all_names
}
\description{
A parsed list of Latin and Swedish names for most species that are
commonly found in Sweden. This list have been extracted from the source
using the scripts available in the repo of this package.
}
\keyword{datasets}
