% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\docType{class}
\name{Analysis-class}
\alias{Analysis-class}
\title{Analysis
analysis of various experiment}
\value{
Analysis
}
\description{
Analysis
analysis of various experiment
}
\section{Slots}{

\describe{
\item{\code{project}}{character. project name.}

\item{\code{author}}{character. author name.}

\item{\code{analysis_name}}{character. analysis name.}

\item{\code{experiment}}{BulkExperiment. experiment need to be analyzed.}

\item{\code{output}}{list. output, results put in analysis}

\item{\code{log}}{character. any comment.}
}}

