% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqiCalculation.R
\name{aqiFromPm25}
\alias{aqiFromPm25}
\title{This function takes in one parameters:pm25Concentration.
It then calculates the AQI given that specified value.}
\usage{
aqiFromPm25(pm25Concentration)
}
\arguments{
\item{pm25Concentration}{that you want to know the AQI for}
}
\value{
it returns the AQI that corresponds to the given PM2.5 measurement
}
\description{
This function takes in one parameters:pm25Concentration.
It then calculates the AQI given that specified value.
}
