% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{AsHMS}
\alias{AsHMS}
\title{Convert time to H:M:S format}
\usage{
AsHMS(secs)
}
\arguments{
\item{secs}{time period in seconds}
}
\value{
Time period formatted for SLURM
}
\description{
Convert time to H:M:S format
}
