% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment-details.R
\name{augment_with_parameters}
\alias{augment_with_parameters}
\title{Augment results with definitional parameters.}
\usage{
augment_with_parameters(data, tables = get_tables())
}
\arguments{
\item{data}{a dataframe from the simulation output; typically derived from
one of the output of the \code{load_duckdb()} or \code{get_tables()} functions.}

\item{tables}{the output of the  \code{load_duckdb()} or \code{get_tables()} function}
}
\value{
a dataframe based on \code{data} with additional columns containing
the group definitional parameters if they exist.
}
\description{
Add in definitional parameters for each combination of the model and
experiment run that are contained in the output. Only the parameters that
are different between groups are added here.
}
\examples{
set_results_dir(here("../scratch/examples/lockdown-compliance"))
get_table("summary") \%>\% augment_with_parameters() 
}
