% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg.R
\name{avg}
\alias{avg}
\title{Calculate the average matching rate from the selected matching rate file.}
\usage{
avg(x)
}
\arguments{
\item{x}{a matching rate file.}
}
\description{
Calculate the average matching rate from the selected matching rate file.
}
