% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blasttools.R
\name{BlastParseNCBI}
\alias{BlastParseNCBI}
\title{Parse blast output}
\usage{
BlastParseNCBI(DGEList, blastRes)
}
\arguments{
\item{DGEList}{count data in the form of DGEList}

\item{blastRes}{file with blast results assumes blastoutput option
-outfmt "6 qseqid sseqid pident length mismatch gapopen qstart qend
sstart send evalue bitscore staxids sscinames scomnames qcovs"}
}
\value{
seqTax a dataframe with annotation of results from sequence
comparisons using blast.
}
\description{
This function will create a dataframe with sequence names, sequence
similarity from blast analysis together with taxonomic information
from NCBI. It needs the blast analysis result file locally
available. This needs to contain the content specified below in the
parameter blastRes. Check the example file test.out included with the
package to see the requirements of the input file.
}
\details{
Needs an active internet connection that can reach NCBI to work.
NB! for optimal performance make sure to get a token from NCBI.
This function is not needed if you have version 1.0 of the
MetAlPipe package.
}
\examples{
\dontrun{
filteredDataFolder <- system.file("extdata", package = "MetAlPipe")
exampleCollectData <- CollectData(filteredDataFolder)
parseTest <- DadaAnalysis(exampleCollectData, muThread = FALSE)
dfForward <- as.data.frame(t(parseTest))
yForward <- edgeR::DGEList(dfForward)
exOut <- system.file("extdata", "test.out", package = "MetAlPipe")
BlastParseNCBI(DGEList = yForward, blastRes = exOut)
}
}
