% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blasttools.R
\name{BlastResWriter}
\alias{BlastResWriter}
\title{Save MultiBlaster result as suitable for BlastParse analysis}
\usage{
BlastResWriter(inputDF, outName)
}
\arguments{
\item{inputDF}{dataframe created with MultiBlaster}

\item{outName}{name of file to write to}
}
\value{
NULL saves a file in current workning directory
}
\description{
Writes at tab delimited file to current working directory with all
information needed to parse the file using the BlastParse
function.
}
\examples{
\dontrun{
Fastafile <- system.file("extdata", "test.fa", package = "MetAlPipe")
inputDF <- MultiBlaster(fastaFile = Fastafile)
BlastResWriter(inputDF = inputDF, outName = "MetAlPipeTest.out")
}
}
