% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{bootstrap_means}
\alias{bootstrap_means}
\title{Calculate means from bootstrap samples}
\usage{
bootstrap_means(mat)
}
\arguments{
\item{mat}{A matrix where each column is one bootstrapped sample}
}
\value{
A vector containing the mean for each boostrapped sample
}
\description{
Calculates a vector of means from a matrix of boostrap samples
}
\examples{
bootstrap_means(mat)
}
