% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{bootstrap_sample}
\alias{bootstrap_sample}
\title{Take bootstrap samples from a vector}
\usage{
bootstrap_sample(vec, num_samples)
}
\arguments{
\item{vec}{The vector you want to sample from (often a column from a dataframe)}

\item{num_samples}{The number of bootstrap samples you want to take}
}
\value{
A matrix of num_samples columns each containing a resampled dataset
}
\description{
Creates a matrix containing bootstrap samples from your original data
}
\examples{
bootstrap_sample(mtcars$mpg, 15000)
}
