% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{BurnOff}
\alias{BurnOff}
\title{Remove burnin}
\usage{
BurnOff(values, burnin)
}
\arguments{
\item{values}{series of parameter estimates}

\item{burnin}{fraction or number of samples to omit from start}
}
\value{
\verb{BurnOff() returns }values\verb{, without the first }burnin` samples
}
\description{
Remove burnin
}
