% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{calculate_model_performance}
\alias{calculate_model_performance}
\title{Calculate hindcast performance for a model}
\usage{
calculate_model_performance(data, model, years)
}
\arguments{
\item{data}{(data.frame) The data to use}

\item{model}{(character) The model formula}

\item{years}{(numeric) Which years to hindcast}
}
\value{
(data.frame) Summarized model performance
}
\description{
Calculate hindcast performance for a model
}
