% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{calculate_model_performance_single}
\alias{calculate_model_performance_single}
\title{Calculate hindcast performance of a model for a single year}
\usage{
calculate_model_performance_single(data, model, year, round_method = floor)
}
\arguments{
\item{data}{(data.frame) The data to use}

\item{model}{(character) The model formula}

\item{year}{(numeric) Which year to forecast}

\item{round_method}{(function) A function to apply to each prediction
(numeric)}
}
\value{
(data.frame) Model performance for a single year
}
\description{
Calculate hindcast performance of a model for a single year
}
