% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{check_categorical}
\alias{check_categorical}
\title{Identifies whether or not a variable is categorical}
\usage{
check_categorical(vec)
}
\arguments{
\item{vec}{A vector}
}
\value{
TRUE/FALSE depending on whether or not the variable is character or factor type
}
\description{
Returns TRUE if a variable is either a character type or a factor type
}
