% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearSecretsWithSubstring.R
\name{clearSecretsWithSubstring}
\alias{clearSecretsWithSubstring}
\title{Clear any key-value pair where the pair contains the substring, ss.}
\usage{
clearSecretsWithSubstring(ss, vault)
}
\value{
The vault
}
\description{
Clear any key-value pair where the pair contains the substring, ss.
}
\examples{

}
