% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CollectData}
\alias{CollectData}
\title{Collect sample names and paths}
\usage{
CollectData(directory = "../Filtered_data", prefix = "")
}
\arguments{
\item{directory}{path to directory with filtered data}

\item{prefix}{only files with the chosen prefix will be identified, default is "".}
}
\value{
list with sample paths and sample names
}
\description{
Collect sample names and paths
}
\examples{
filteredDataFolder <- system.file("extdata", package = "MetAlPipe")
CollectData(filteredDataFolder)

}
