% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CombineSpecies}
\alias{CombineSpecies}
\title{CombineSpecies}
\usage{
CombineSpecies(ref1, ref2, countsObject)
}
\arguments{
\item{ref1}{a string containing the name of the first species to be combined (this name will be kept).}

\item{ref2}{a string containing the name of the second species to be combined.}

\item{countsObject}{a dataframe containing counts of species with the species names in the first column to the left.}
}
\value{
The input dataframe with the two species counts for ref1 and ref2 combined under the ref1 name.
}
\description{
Takes the names of two species and a dataframe of counts and returns the dataframe with the counts of the two species combined.
}
