% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{compactNameDump}
\alias{compactNameDump}
\title{Taxonomic names from NCBI}
\format{
\subsection{\code{compactNameDump}}{

A data frame with 2491920 rows and 2 columns:
\describe{
\item{V1}{Node number}
\item{V2}{Taxonomic name}
}
}
}
\source{
\url{https://ftp.ncbi.nih.gov/pub/taxonomy/taxdmp.zip}
}
\usage{
compactNameDump
}
\description{
A parsed list of Taxonomic names from NCBI. This list have been extracted from the source
using scripts and information available in the repo of this package.
}
\keyword{datasets}
