% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{compactNodeDump}
\alias{compactNodeDump}
\title{Taxonomic ranks related to nodes in the compactNameDump}
\format{
\subsection{\code{compactNodeDump}}{

A data frame with 2491920 rows and 3 columns:
\describe{
\item{V1}{Node numbers}
\item{V2}{Node numbers}
\item{V3}{Taxonomic levels}
}
}
}
\source{
\url{https://ftp.ncbi.nih.gov/pub/taxonomy/taxdmp.zip}
}
\usage{
compactNodeDump
}
\description{
A parsed list of Taxonomic names from NCBI. This list have been extracted from the source
using scripts and information available in the repo of this package.
}
\keyword{datasets}
