% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dataframes.R
\name{compareDataDF}
\alias{compareDataDF}
\title{This function takes in a data frame of values with columns timestamp,
otherCityPM, southGatePM, and PM2.5. It also takes in the name of the city
returns the results after conducting t-tests on them}
\usage{
compareDataDF(ourData, nameOfCity)
}
\arguments{
\item{two}{arrays of data with matching days and the name of the non-South Gate city}
}
\value{
it returns the t test results of the data sets
}
\description{
This function takes in a data frame of values with columns timestamp,
otherCityPM, southGatePM, and PM2.5. It also takes in the name of the city
returns the results after conducting t-tests on them
}
