% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeComparison.R
\name{ComparisonTrees}
\alias{ComparisonTrees}
\title{Sample of 128 posterior trees for comparison with well-corroborated tree}
\usage{
ComparisonTrees(scriptID, pID)
}
\arguments{
\item{scriptID}{Character specifying the script to run}

\item{pID}{Character specifying the project identifier to analyse}
}
\description{
\code{ComparisonTrees} samples trees from the posterior; then drops and relabel
leaves until the taxon set matches that in the corresponding well
corroborated trees
}
