% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{Config}
\alias{Config}
\title{Configuration settings}
\usage{
Config()
}
\description{
View and edit package configuration variables
}
\examples{
ntConfig <- Config()

# View maximum cores that will be requested via SLURM
.config$maxCores
  
\dontrun{
  # Override default matrix directory
  ntConfig$matrixDir <- "path/to/my/matrices" 
}


}
