% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{create_performance_table}
\alias{create_performance_table}
\title{Create a model performance table}
\usage{
create_performance_table(data, models, years)
}
\arguments{
\item{data}{(data.frame) The data to use}

\item{models}{(character) The model formula}

\item{years}{(numeric) Which years to hindcast}
}
\value{
(data.frame) A model performance table
}
\description{
Create a model performance table
}
