% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CurateHit}
\alias{CurateHit}
\title{Curate hits (wrapper function)}
\usage{
CurateHit(
  ref,
  note = "",
  absorb = "",
  latin = "",
  common = "",
  rmHit = FALSE,
  countsObject
)
}
\arguments{
\item{ref}{the reference of the sequence to curate.}

\item{note}{a string containing the note to be added (no action is taken if string is empty).}

\item{absorb}{a string or vector specifying the reference(s) of another hit(s) to absorb / be combined with (no action is taken if string is empty).}

\item{latin}{a string specifying new latin name (no action is taken if string is empty).}

\item{common}{a string specifying new common name (no action is taken if string is empty).}

\item{rmHit}{boolean specifying if the hit should be removed or not (default: FALSE).}

\item{countsObject}{dataframe with columns 1:5 with sequence info (unique sequence references, notes, Swedish name, Latin name and percentage out of all sequences in the data frame) and the rest of the columns of samples with counts of the sequences.}
}
\value{
dataframe with columns 1:5 with sequence info (sequence reference, notes, Swedish name, Latin name and percentage out of all sequences in the data frame) and the rest of the columns of samples with counts of the sequences.
}
\description{
Curate hits (wrapper function)
}
