% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dataframes.R
\name{dailySG}
\alias{dailySG}
\title{Daily Summary Data for South Gate}
\usage{
dailySG(data)
}
\arguments{
\item{data}{the output dataframe of the hourlyPA or cleanPA functions}
}
\value{
a dataframe of the holistic South Gate data, including \emph{weighted} sensor averages, sensor median, the max, and the min, and active count.
}
\description{
This function aggregates data (output from the hourlyPA/cleanPA function) by hour for mean, median, max, min, and count of sensors
It assumes that the data has the columns 'timestamp', 'day', 'hour' 'PM2.5.
}
