% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{deltaAnalysis}
\alias{deltaAnalysis}
\title{deltaAnalysis}
\usage{
deltaAnalysis(
  assay,
  scale = F,
  assay_name = "assay",
  sample_info,
  patient_id_column = "Patient_ID",
  sample_id_column = "Sample_ID",
  timepoint_column,
  timepoint_terms = c(baseline = "Baseline", tp2 = "TP2"),
  group_column,
  top_anno,
  top_anno_patient_order,
  test_method = c("ttest", "limma"),
  contrasts,
  pval_cutoff = 0.05,
  only_sig_genes = F,
  only_show_sig_genes = F,
  palette,
  output_dir
)
}
\arguments{
\item{assay}{data.frame. assay data.}

\item{scale}{logic. whether to scale assay.}

\item{assay_name}{character. assay name.}

\item{sample_info}{data.frame. sample information.}

\item{patient_id_column}{character. column for patient id.}

\item{sample_id_column}{character. column for sample id.}

\item{timepoint_column}{character. column for timepoint.}

\item{timepoint_terms}{character. timepoint terms pair as delta.}

\item{group_column}{character. column for group.}

\item{top_anno}{HeatmapAnnotation. top annotation, passed to heatmap.}

\item{top_anno_patient_order}{character. patient orders for top annotation.}

\item{test_method}{character. test method, such as "ttest" or "limma".}

\item{contrasts}{character. specific contrast.}

\item{pval_cutoff}{numeric. p value cutoff for significance.}

\item{only_sig_genes}{logic. whether only significant genes will be used.}

\item{only_show_sig_genes}{logic. whether only significant genes will be showed in heatmap.}

\item{palette}{character. colors for KM curve.}

\item{output_dir}{character. output directory.}
}
\value{
list. contain combined_assay,combined_assay_heapmap,combined_statistics,group_terms,groups.
}
\description{
delta assay anlysis
}
