% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsherSmithSetup.R
\name{DeZZ}
\alias{DeZZ}
\title{Unify tip label format}
\usage{
DeZZ(tr)
}
\arguments{
\item{tr}{tree whose labels should be rewritten}
}
\value{
\code{DeZZ()} returns a tree whose labels have been rewritten in a
consistent format, with underscores replaecd with spaces and leading \code{zz}
prefixes removed.
}
\description{
Unify tip label format
}
