% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DFCombine}
\alias{DFCombine}
\title{Combine forward and reverse runs if reverse runs are present}
\usage{
DFCombine(dataset, samples)
}
\arguments{
\item{dataset}{path to directory with filtered data}

\item{samples}{vector with sample names}
}
\value{
dataframe with the number reads in every sample
}
\description{
Combine forward and reverse runs if reverse runs are present
}
