% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{difAnalysis}
\alias{difAnalysis}
\title{difAnalysis}
\usage{
difAnalysis(
  assay,
  scale = F,
  assay_name = "assay",
  sample_info,
  patient_id_column = "Patient_ID",
  sample_id_column = "Sample_ID",
  group_column,
  block_column = NULL,
  contrasts = NULL,
  top_anno,
  top_anno_sample_order,
  test_method = c("ttest", "limma"),
  pval_cutoff = 0.05,
  only_sig_genes = F,
  palette = NULL,
  output_dir,
  ...
)
}
\arguments{
\item{assay}{data.frame. assay data.}

\item{scale}{logic. whether to scale assay.}

\item{assay_name}{character. assay name.}

\item{sample_info}{data.frame. sample information.}

\item{patient_id_column}{character. column for patient id.}

\item{sample_id_column}{character. column for sample id.}

\item{group_column}{character. column for group.}

\item{block_column}{character. column for block factor.}

\item{contrasts}{character. specific contrast.}

\item{top_anno}{HeatmapAnnotation. top annotation, passed to heatmap.}

\item{top_anno_sample_order}{character. sample orders for top annotation.}

\item{test_method}{character. test method, such as "ttest" or "limma".}

\item{pval_cutoff}{numeric. p value cutoff for significance.}

\item{only_sig_genes}{logic. whether only significant genes will be used.}

\item{palette}{character. colors used for heatmap.}

\item{output_dir}{character. output directory.}

\item{...}{list. parameter passed tp ComplexHeatmap::Heatmap}
}
\value{
list. contain various contrast result.
}
\description{
differential assay anlysis
}
