% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{DiffDepth}
\alias{DiffDepth}
\title{Explore whether differences in tree topology are deep or shallow}
\usage{
DiffDepth(pID, models, pdf = FALSE, sampleSize = 128, resamplings = 1024 * 4)
}
\description{
Explore whether differences in tree topology are deep or shallow
}
