% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{This function takes in four parameters: lat_1,long_1,lat_2,long_2.
It then calculates the distance between these two locations using
the Haversine formula.}
\usage{
distance(lat_1, long_1, lat_2, long_2)
}
\arguments{
\item{latitude}{of the first location, longitude of the first location, latitude
of the second location, longitude of the second location}
}
\value{
it returns the distance between these locations in miles
}
\description{
This function takes in four parameters: lat_1,long_1,lat_2,long_2.
It then calculates the distance between these two locations using
the Haversine formula.
}
