% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{.DiffErr}
\alias{.DiffErr}
\title{Combine two error terms}
\usage{
.DiffErr(err1, err2)
}
\arguments{
\item{err1, err2}{Numeric vectors giving errors to be combined.}
}
\value{
\code{.DiffErr()} returns a vector detailing the combined error for
each element in \code{err1}.
}
\description{
The error associated with a difference can be computed from the error
associated with each input variable
}
