% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-standalone-ggplot-utils.R
\name{.gg_scale_y_percent}
\alias{.gg_scale_y_percent}
\title{A percentage y scale}
\usage{
.gg_scale_y_percent(..., sf = 2)
}
\arguments{
\item{...}{Other arguments passed on to \verb{scale_(x|y)_continuous()}}

\item{sf}{significant figures}
}
\value{
a ggplot scale
}
\description{
A percentage y scale
}
\concept{ggplot}
\keyword{internal}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{


tibble::tibble(pvalue = c(0.001, 0.05, 0.1), fold_change = 1:3) %>%
 ggplot2::ggplot(ggplot2::aes(fold_change , pvalue)) +
 ggplot2::geom_point() +
 scale_y_percent()
}\if{html}{\out{</div>}}
}

