% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-standalone-directory-utils.R
\name{.inputter}
\alias{.inputter}
\title{Find a file in a data directory.}
\usage{
.inputter(directory = .here("input"))
}
\arguments{
\item{directory}{the root of the input}
}
\value{
a function that takes a relative path and returns the absolute path
of the input file. The function can take the same arguments as \code{fs::dir_ls},
and particularly useful is \code{glob} and \code{type}. The function
}
\description{
This function generates a function that resolves a file path fragment to a
specific file location in an input directory.
}
\concept{output}
\keyword{internal}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{

inp = .inputter(tempdir())
for (i in 1:10) {
  fs::file_touch(fs::path(tempdir(), sprintf("test_%d.txt",i)))
}

testthat::expect_true(fs::file_exists(inp("test_2.txt")))
testthat::expect_error(inp("test_0.txt"))

testthat::expect_equal(length(inp(glob="*.txt")),10)

}\if{html}{\out{</div>}}
}

