% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFuncsImp.R
\name{.MeanAcc}
\alias{.MeanAcc}
\title{Mean accuracy across all tips and characters for one replicate}
\usage{
.MeanAcc(pooled)
}
\description{
Mean accuracy across all tips and characters for one replicate
}
