% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{.NCoded}
\alias{.NCoded}
\title{Number of characters with non-ambiguous state}
\usage{
.NCoded(path)
}
\arguments{
\item{path}{Path to nexus file}
}
\description{
Number of characters with non-ambiguous state
}
