% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{.OutwithError}
\alias{.OutwithError}
\title{Outwith error}
\usage{
.OutwithError(x, err, eps = epsBF)
}
\arguments{
\item{x}{Value to compare; typically a Bayes Factor giving difference
in estimated marginal likelihoods of two models}

\item{err}{Standard error of comparison, perhaps computed using \code{.DiffErr()}}

\item{eps}{Numeric; errors smaller than \code{eps} are not deemed noteworthy.}
}
\value{
\code{.OutwithError()} returns a logical vector stating whether each
element in \code{x} is greather than the corresponding \code{eps} and \code{err}
}
\description{
Do two estimates differ by a noteworthy amount?
}
