% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-standalone-directory-utils.R
\name{.this_script}
\alias{.this_script}
\title{Get the file path of the current script}
\usage{
.this_script()
}
\value{
the file path of the currently executed script or an error if the
command is executed outside of a script.
}
\description{
Gives you the file input path regardless of whether you are running the
script in rstudio, knitr or on the console.
}
\concept{output}
\keyword{internal}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{

.this_script()
}\if{html}{\out{</div>}}
}

