% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{EpsLine}
\alias{EpsLine}
\title{Mark meaningful differences}
\usage{
EpsLine(horiz = TRUE, vert = FALSE, diag = FALSE)
}
\description{
Adds lines to plots at position of \code{.config$eps}, i.e. the smallest
likelihood difference deemed to support one model over another.
}
