% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{EvaluatePartitioning}
\alias{EvaluatePartitioning}
\title{Evaluate t parameter}
\usage{
EvaluatePartitioning(marginals, suffix = "_ki")
}
\description{
Does setting a free \eqn{t} parameter improve model fit relative to a fixed
\eqn{t = 1}?
}
