% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckComplete.R
\name{ExistingResults}
\alias{ExistingResults}
\title{Existing results
Read results that have already been cached using \code{UpdateRecords()}.}
\usage{
ExistingResults(pID, scriptID, checkRemote = FALSE)
}
\arguments{
\item{pID}{Character specifying the project identifier to analyse}

\item{scriptID}{Character specifying the script to run}

\item{`checkRemote`}{Logical. If results do not exist locally, should we
attempt to retrieve them from the remote server?}
}
\description{
Existing results
Read results that have already been cached using \code{UpdateRecords()}.
}
