% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckComplete.R
\name{ExtendRun}
\alias{ExtendRun}
\title{Submit a continuation job}
\usage{
ExtendRun(pID, scriptID, proportion, buffer = 1.15)
}
\arguments{
\item{proportion}{How much longer should the run be?
e.g. if the observed ess is 100 and we need an ESS of 300, proportion = 3}

\item{buffer}{Ratio to increase proportion to be sure we clear
our threshold.  1.15 is a 15\% buffer.}
}
\description{
Submit a continuation job
}
