% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{extract_categoricals}
\alias{extract_categoricals}
\title{Extract categorical variables from a data frame}
\usage{
extract_categoricals(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
A new dataframe with just categorical variables
}
\description{
Returns a new dataframe consisting only of the categorical variables from the original data frame
}
