% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{extract_numerics}
\alias{extract_numerics}
\title{Extract numeric variables from a data frame}
\usage{
extract_numerics(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
A new dataframe with only numeric variables
}
\description{
Returns a new data frame consisting only of numeric variables
}
