% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFuncs.R
\name{FetchLogIfMissing}
\alias{FetchLogIfMissing}
\title{Fetch log file from simulation study}
\usage{
FetchLogIfMissing(seed, logFile)
}
\description{
set \code{options("requeue" = TRUE)} to automatically re-execute any inference
for which a log file is missing
}
