% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{FetchResults}
\alias{FetchResults}
\title{Fetch results from remote server}
\usage{
FetchResults(pID, scriptID)
}
\arguments{
\item{pID}{Character specifying the project identifier to analyse}

\item{scriptID}{Character specifying the script to run}
}
\description{
Fetch results from remote server
}
