% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{FiltTrimWrap}
\alias{FiltTrimWrap}
\title{Convenience function to run dada2}
\usage{
FiltTrimWrap(primerData)
}
\arguments{
\item{primerData}{list containing information about samples to be run, can be generated with the function \link{CollectData}.}
}
\value{
summary of sequence data as described in the Dada2 package
}
\description{
This function filters read with low quality or containing N
with default parameters.
}
