% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{FractionPlot}
\alias{FractionPlot}
\title{Barplots with fractions}
\usage{
FractionPlot(data, species, col, inset = c(-0.3, 0.01))
}
\arguments{
\item{data}{a dataframe with count data from each species on rows}

\item{species}{a vector with names of the species in the same order
as the appear in the data}

\item{col}{of colors to use in legend and plot.}

\item{inset}{inset values for legend plot}
}
\value{
a barplot
}
\description{
Creates a simple horizontal barplot showing the fraction of the
total from each species.
}
\examples{
data <- data.frame(A = 1:3, B = c(1,4,0), C = c(2, 3, 7))
species <- c("Abramis brama", "Salmo trutta", "Salmo salar")
col <- c("blue", "orange", "grey")
inset <- c(-0.3, 0.01)
FractionPlot(data = data, species = species, col = col, inset = inset)

}
