% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{generate_quantiles_normal}
\alias{generate_quantiles_normal}
\title{Generates normal quantiles from a sample set}
\usage{
generate_quantiles_normal(vec)
}
\arguments{
\item{vec}{A vector containing your actual data}
}
\value{
A vector containing the sample normal quantiles
}
\description{
Generates a set of normal quantiles to compare to your data to check for normality
}
