% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GenerateASV}
\alias{GenerateASV}
\title{Generate Amplicon Sequence Variant Table}
\usage{
GenerateASV(
  blastObject = blastResY,
  countsObject = yAll,
  classFilter = c("Actinopteri", "Hyperoartia", "Chondrichthyes"),
  filterByClass = FALSE
)
}
\arguments{
\item{blastObject}{object containing the BLAST results}

\item{countsObject}{object containing sequences and counts}

\item{classFilter}{vector of class hits to keep, or simply a string}

\item{filterByClass}{boolean if class filter should be applied or not}
}
\value{
ASV table in the form of a data.frame
}
\description{
Generate Amplicon Sequence Variant Table
}
