% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-experiments.R
\name{geom_policy}
\alias{geom_policy}
\title{Plot a policy background layer.}
\usage{
geom_policy(alpha = FALSE)
}
\arguments{
\item{alpha}{the default plot is stacked proportions of a particular policy
which would work well as a rug plot underneath a time series. Alternatively
if alpha is \code{TRUE} the policy state is shown as a solid background with
varying hues and degrees of alpha. This is useful for the background of
spaghetti plots of incidence, but this will only be displayed once another
layer is present. In both cases a sensible fill scale needs to be chosen
that works well with discrete data.}

\item{palette}{the palette as a function which takes a number of levels and
returns a colour for that level. In all cases the default (i.e. time zero)
policy will be transparent white so changes will only be obvious for
policies that are a change from the baseline.}
}
\value{
a ggplot timeseries layer with the background layer holding the policy
state over time. If \code{alpha=TRUE} this will only display a meaningful output
if some other layer is providing the y axis.
}
\description{
This plots the simulation wide policy in place at any given point in time
in the simulation(s). Multiple simulations may have different policies at
different time (e.g. lockdowns will start and end at different times in
different simulations) These can be represented either by a gradual colour
transition in the background (\code{alpha=TRUE}) or as a stacked bar chart of the
number of simulations in a given policy state (\code{alpha=FALSE})
}
\examples{
set_results_dir(here("../scratch/examples/lockdown-compliance"))
ggplot2::ggplot()+geom_policy(alpha = TRUE)+ggplot2::geom_hline(yintercept=1)

}
