% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment-details.R
\name{get_experiment_settings}
\alias{get_experiment_settings}
\title{Read experiment result settings}
\usage{
get_experiment_settings(dir = get_results_dir())
}
\arguments{
\item{dir}{Character. Path to the results directory. Defaults to \code{get_results_dir()}.}
}
\value{
A named list containing the parsed experiment settings (result of \code{jsonlite::fromJSON}).
}
\description{
Read and parse the experiment settings stored as JSON in the results directory.
If the directory is not provided then it is
}
\examples{
\dontrun{
settings <- get_experiment_settings()        # use default results directory
settings <- get_experiment_settings("out/1") # specify a directory
}
}
