% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{get_policy_colours}
\alias{get_policy_colours}
\title{Get the colours and names of policy state highlights.}
\usage{
get_policy_colours()
}
\value{
a names list
}
\description{
Get the colours and names of policy state highlights.
}
