% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{get_results_dir}
\alias{get_results_dir}
\title{Get the JPanSim2 output directory.}
\usage{
get_results_dir()
}
\value{
the cached directory path. If missing an interactive folder chooser
will be displayed.
}
\description{
Get the JPanSim2 output directory.
}
\examples{
set_results_directory(tempdir())
get_results_directory()
}
