% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{get_table}
\alias{get_table}
\title{Get the JPanSim2 output directory.}
\usage{
get_table(name)
}
\value{
a single duckdb table
}
\description{
Get the JPanSim2 output directory.
}
\examples{
set_results_dir(tempdir())
get_tables()
}
