% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getRawSeqs}
\alias{getRawSeqs}
\title{Collect raw fragment counts from FastQC output}
\usage{
getRawSeqs(searchFolder = "../Pre_analysis", searchPattern = "1_fastqc.html")
}
\arguments{
\item{searchFolder}{path to folder with FastQC output files}

\item{searchPattern}{file ending for forward reads}
}
\value{
dataframe with fragment counts
}
\description{
Collect raw fragment counts from FastQC output
}
