% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetTaxonomy}
\alias{GetTaxonomy}
\title{Get taxonomic information from NCBI}
\usage{
GetTaxonomy(searchName, nameDump = compactNameDump, nodeDump = compactNodeDump)
}
\arguments{
\item{searchName}{string of species name to extract taxonomic information about}

\item{nameDump}{dataframe with names and node information}

\item{nodeDump}{dataframe with names and node information}
}
\value{
vector with taxonomic information
}
\description{
Takes a species name and extract taxonomic information from compactNameDump
and compactNodeDump available in the MetAlPipe package.
}
\details{
Attach data with:
data(compactNameDump, package = "MetAlPipe")
data(compactNodeDump, package = "MetAlPipe")
}
\examples{
GetTaxonomy(searchName = "Esox lucius",
            nameDump = compactNameDump,
            nodeDump = compactNodeDump)

}
