% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettingCityName.R
\name{gettingCityName}
\alias{gettingCityName}
\title{This function takes in a read csv file from AQMD and returns the
name of the city that the data is about}
\usage{
gettingCityName(data)
}
\arguments{
\item{a}{csv file}
}
\value{
it returns the name that the data set is about.
}
\description{
This function takes in a read csv file from AQMD and returns the
name of the city that the data is about
}
