% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment-details.R
\name{group_by_simulation_day}
\alias{group_by_simulation_day}
\title{Group the output of the results by simulation run and day}
\usage{
group_by_simulation_day(data, ..., replicas = FALSE)
}
\arguments{
\item{data}{a dataframe from the simulation output; typically derived from
one of the output of the \code{load_duckdb()} or \code{get_tables()} functions.}

\item{...}{additional columns in \code{data} to group by over and above
the simulation identifiers}

\item{replicas}{treat each replica of a simulation separately? defaults to \code{FALSE}}
}
\value{
a grouped dataframe using default columns where each group is an
single day in one simulation run (\code{replicas=TRUE}) or a single day in all
simulation replicas (\code{replicas=TRUE})
}
\description{
Group the output of the results by simulation run and day
}
