% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{group_deltas}
\alias{group_deltas}
\title{Get preserved group characteristics that define groups}
\usage{
group_deltas(grp_df)
}
\arguments{
\item{grp_df}{a grouped data frame}
}
\value{
a dataframe that contains the groups and other variables that do not
vary within group but do vary between groups, plus other distractors that
either vary within group or are constant over all groups.
}
\description{
Group characteristics that define a group are the same within a group but
vary between groups so in some regard define the group. This function returns
the data frame grouped with only those defining characteristics preserved. It
will enter nested dataframes as well.
}
\examples{
grp_df = ggplot2::diamonds \%>\% 
  dplyr::mutate(idx = as.numeric(cut) * as.numeric(color), dup=1) \%>\% 
  dplyr::group_by(cut, color) 
grp_df \%>\% group_deltas()

# In our case:
tables$parameters \%>\% 
  dplyr::group_by(experimentName, modelName) \%>\% 
  group_deltas()
}
