% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{HasConverged}
\alias{HasConverged}
\title{Has an analysis converged?}
\usage{
HasConverged(
  pID,
  scriptID,
  pt = .config$psrfThreshold,
  et = .config$essThreshold
)
}
\arguments{
\item{pID}{Character specifying the project identifier to analyse}

\item{scriptID}{Character specifying the script to run}

\item{pt}{Gelman-Rubin statistic (potential scale reduction factor) threshold;
analyses with PSRF > \code{pt} have not converged.}

\item{et}{Estimated sample size threshold; analyses with ESS < \code{et} have
not converged.}
}
\value{
\code{HasConverged()} returns a logical specifying whether the specified
analysis has converged at the specified thresholds.
}
\description{
Has an analysis converged?
}
