% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{here}
\alias{here}
\title{Drop in replacement for \code{here} (\code{here} pkg)}
\usage{
here(..., projRoot = .locate_project())
}
\arguments{
\item{...}{the relative path within the project}

\item{projRoot}{the project root - defaults to \code{.locate_project()}}
}
\value{
a path
}
\description{
Drop in replacement for \code{here} (\code{here} pkg)
}
\examples{
print(here())
}
