% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{HetNPlot}
\alias{HetNPlot}
\title{Plot values of \emph{n} from heterogeneous models}
\usage{
HetNPlot(hgN, hgModelBF, hgModel)
}
\arguments{
\item{hgN}{Numeric matrix with named rows giving 2.5, 25, 50, 75 and 97.5\%
quantiles for mean N value for each dataset}

\item{hgModelBF}{Numeric giving Bayes factor associated with each model,
compared to Mk model}

\item{hgModel}{character specifying \code{scriptID} corresponding to the chosen
model.}
}
\description{
Plot values of \emph{n} from heterogeneous models
}
