% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dataframes.R
\name{hourlyCleanPA}
\alias{hourlyCleanPA}
\title{This function cleans up the raw readings from the PA sensors and sets up a data frame.
It assumes that the data has the columns
c('timestamp', "channelAPm25", "channelBPm25", "humidity", "latitude", "longitude") in that order.
It also uses an EPA correction factor for PurpleAir sensors.
Any readings with faulty channel A/B values will be dropped}
\usage{
hourlyCleanPA(test1)
}
\arguments{
\item{data}{a .csv of PurpleAir sensor data from the CEHAT website}
}
\value{
a dataframe of the full PurpleAir sensor data (by the hour), with corrected timezone and combined channels
}
\description{
This function cleans up the raw readings from the PA sensors and sets up a data frame.
It assumes that the data has the columns
c('timestamp', "channelAPm25", "channelBPm25", "humidity", "latitude", "longitude") in that order.
It also uses an EPA correction factor for PurpleAir sensors.
Any readings with faulty channel A/B values will be dropped
}
