% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ImportMeta}
\alias{ImportMeta}
\title{Read and make project metadata available in analysis}
\usage{
ImportMeta(file)
}
\arguments{
\item{file}{excel file that holds the metadatat as follows:
Project    - Internal project number (character)
Client     - Organisation that placed the order (character)
TaxGroup   - Taxonmomic group that is studied (Valid taxonomic group)
Marker     - Name and length of marker used (named vector)
Primers    - Primer sequence used (character vector)
SampleName - Sample names (character vector)
GISdata    - GPS coordinates for samples in WGS84
Comments   - Optional (character vector)
SequenceC  - Novoseq or Scilifelab currently the only supported
Method     - MiSeq, NanoSeq, NovaSeq the currently the only supported
ReadLength - Single read sequence length
Pair-end   - TRUE/FALSE}
}
\description{
Creates a list of entries with metadata that can be used in report
and analysis.
}
