% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnDiscretizeBeta.R
\name{incompleteBeta}
\alias{incompleteBeta}
\title{Discretized beta distribution
R implementation of the RevBayes function fnDiscretizeBeta}
\source{
https://gist.githubusercontent.com/ms609/883632d10d4d80ea5391cee9c47071fc/raw/505b31f322060dd2cdde8010ae74708751ee862e/fnDiscretizeBeta.R
Martin R. Smith
2024-12-05
}
\usage{
incompleteBeta(a, b, x)
}
\description{
Discretized beta distribution
R implementation of the RevBayes function fnDiscretizeBeta
}
