\name{indexCalculation}
\alias{indexCalculation}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Calculating AQI given Indices}
\description{
This function takes in five parameters: highIndexBreakpoint, lowIndexBreakpoint, highConcentrationBreakpoint, lowConcentrationBreakpoint,pm25Concentration, in
that order. It then calculates the AQI given the specified indices.
}
\usage{
indexCalculation(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
the function takes in highIndexBreakpoint, lowIndexBreakpoint, highConcentrationBreakpoint, lowConcentrationBreakpoint,pm25Concentration
in that order
}
}
\value{an AQI value given the specifications}
