% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{invariants}
\alias{invariants}
\title{Get preserved group characteristics.}
\usage{
invariants(grp_df, distinct = TRUE)
}
\arguments{
\item{grp_df}{a grouped data frame}

\item{distinct}{defaults to \code{TRUE}, creates a unique row for every group.}
}
\value{
a dataframe that contains the groups and other variables that do not
vary within group.
}
\description{
These are variables that are the same for all members of a group and are
thus effectively a group wide characteristic.
}
\examples{
grp_df = ggplot2::diamonds \%>\% 
  dplyr::mutate(idx = as.numeric(cut) * as.numeric(color), dup=1) \%>\% 
  dplyr::group_by(cut, color) 
grp_df \%>\% invariants()
}
