% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{keys}
\alias{keys}
\title{Return a list of keys currently in the vault}
\usage{
keys(x)
}
\value{
A list of keys currently in the vault.
}
\description{
Return a list of keys currently in the vault
}
\examples{
keys <- keys()
}
