% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilePaths.R
\name{KiProjects}
\alias{KiProjects}
\title{List projects suitable for _ki analyses}
\usage{
KiProjects()
}
\value{
Character vector of project identifiers.
}
\description{
Return identifiers for projects retained for downstream \verb{_ki}
analyses, excluding those known to fail or duplicate others.
}
