% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment-details.R
\name{load_duckdb}
\alias{load_duckdb}
\title{Load all \code{duckdb} files and CSV files}
\usage{
load_duckdb(dir = get_results_dir())
}
\arguments{
\item{dir}{the directory where simulation results are found. This may
be a flat directory from simulations run on a single node, or a set of
directories with the output of multiple nodes. Within the output
directory/ies the individual \code{duckdb} and \code{csv} files are all consistently
named by the simulation.}
}
\value{
A named list:
\itemize{
\item \code{dimensions}: a list which details how many different simulations were
run and the breakdown of whether these were replications or comparisons.
\item \code{experiments}: a lazy \code{tbl} dataframe backed by \code{duckdb} which links the
experiment name with the combination of changes made to the default
configuration for that experiment.
\item \code{models}: a table which links the
experiment name with the combination of changes made to the default
configuration for that model.
\item \code{parameters}: a  lazy \code{tbl} dataframes backed by \code{duckdb} which links
each simulation run to the full parametrisation (of both the model setup
and execution). This has all the configuration used for that model run.
\item remaining entries are lazy \code{tbl} dataframes backed by \code{duckdb}. These can
be queried with \code{dplyr} but natively support will be slightly limited by
the \code{duckdb} backed.
}

The return value will have a \code{con} attribute holding the \code{duckdb}
connection.

When using the \code{comparisons} or \code{parameters} tables, native \code{dplyr} or is
needed for the list columns. The \code{duckdb} tables will need to be copied
into the R session with a \code{dplyr::collect()} call.
}
\description{
This function loads all the outputs of a simulation set into a single
\code{duckdb} container and provides access to it as \code{dplyr} tables. It combines
outputs from multiple SLURM nodes if needed. It augments the raw output data
with the names of the facets in each simulation.
}
\examples{
tables = load_duckdb(here("../scratch/examples/behaviour-comparison"))

}
