% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_holidays.R
\name{load_german_holidays}
\alias{load_german_holidays}
\title{Load German Holidays}
\usage{
load_german_holidays(from_year, to_year, location)
}
\arguments{
\item{from_year}{Integer. The start year for which to retrieve holidays.}

\item{to_year}{Integer. The end year for which to retrieve holidays.}
}
\value{
A vector of `Date` objects representing public holidays in Germany, 
filtered to include only globally recognized holidays.
}
\description{
This function retrieves public holidays in Germany for a specified range of years 
using the Nager.Date API and returns a vector of unique dates that are recognized 
as public holidays.
}
\note{
-
}
\examples{
\dontrun{
# Load holidays from 2015 to 2024
german_holidays <- load_german_holidays(2015, 2024)
}

}
