% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{lollipop}
\alias{lollipop}
\title{Lollipop Chart}
\usage{
lollipop(data, sensor)
}
\arguments{
\item{data}{a .csv of PurpleAir high/low sensor data from the CEHAT website}

\item{sensor}{which sensor you want to view}
}
\value{
a lollipop ggplot2 chart of daily high and low values per sensor for a given time period
}
\description{
This function creates a lollipop chart visualization for PurpleAir high/low sensor data.
}
