% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MakeDGEList}
\alias{MakeDGEList}
\title{Create DGEList}
\usage{
MakeDGEList(dataset, primerData)
}
\arguments{
\item{dataset}{matrix with count data. Samples on rows and variants over columns}

\item{primerData}{list containing information about samples to be run, can be generated with the function \link{CollectData}.}
}
\value{
DGEList
}
\description{
Function that creates a DGEList while also combining forward and reverse runs if reverse runs are present.
}
