% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MakeNotes}
\alias{MakeNotes}
\title{Create an object for curating and taking notes about sequences}
\usage{
MakeNotes(countsObject)
}
\arguments{
\item{countsObject}{data frame with one row with names of species and one to several columns with sequence counts of those species.}
}
\value{
dataframe with columns 1:5 with sequence info (sequence reference, notes, Swedish name, Latin name and percentage out of all sequences in the data frame) and the rest of the columns of samples with counts of the sequences.
}
\description{
Create an object for curating and taking notes about sequences
}
