% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckComplete.R
\name{MarginalDiffs}
\alias{MarginalDiffs}
\title{Difference between marginal likelihood estimates obtained by stepping stone
and path sampling analyses.}
\usage{
MarginalDiffs(projects, models)
}
\arguments{
\item{projects}{Character vector listing project identifiers to evaluate.}

\item{models}{Character vector specifying models to evaluate.}
}
\value{
\code{MarginalDiffs()} returns a matrix specifying the difference
(in log units) between the stepping stone and path sampling estimates of the
marginal likelihood.
}
\description{
Difference between marginal likelihood estimates obtained by stepping stone
and path sampling analyses.
}
