% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{ModelBF}
\alias{ModelBF}
\title{Bayes factor support for each model, relative to best fitting}
\usage{
ModelBF(marginals, models)
}
\arguments{
\item{marginals}{Matrix specifying marginal likelihood of each model (rows)
for each dataset (columns)}

\item{models}{vector identifying the rows of \code{marginals} to be considered}
}
\description{
Bayes factor support for each model, relative to best fitting
}
