% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aesthetic.R
\name{ModelCol}
\alias{ModelCol}
\title{Model colour
Return the colour used to plot results of this model}
\usage{
ModelCol(x)
}
\description{
Model colour
Return the colour used to plot results of this model
}
