% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{ModelHeatmap}
\alias{ModelHeatmap}
\title{Heatmap of model comparisons}
\usage{
ModelHeatmap(marginals, stdErr)
}
\arguments{
\item{marginals, stdErr}{Matrix specifying marginal likelihood (\code{marginals})
and standard error (\code{stdErr}) of each analysis.
Rows correspond to models; columns to datasets}
}
\value{
\code{ModelHeatmap()} is called for its side-effect of plotting a model
heatmap.
}
\description{
Heatmap of model comparisons
}
