% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{ModelIsHeterogeneous}
\alias{ModelIsHeterogeneous}
\alias{ModelIsStationary}
\title{Type of model}
\usage{
ModelIsHeterogeneous(scriptID)

ModelIsStationary(scriptID)
}
\arguments{
\item{scriptID}{Character specifying the script to run}
}
\value{
Returns \code{TRUE} if the model is of the specified type;
\code{FALSE} otherwise.
}
\description{
Type of model
}
