% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aesthetic.R
\name{ModelLabel}
\alias{ModelLabel}
\title{Model label
Translates a model script identifier to a human-readable label}
\usage{
ModelLabel(x)
}
\description{
Model label
Translates a model script identifier to a human-readable label
}
