% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{my_model}
\alias{my_model}
\title{Fit model using any of the following algorithms: logistic regression, tree, gradient boosting, svm}
\usage{
my_model(task, train_set, classifier_name)
}
\description{
This function is not exported
}
