% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dataframes.R
\name{newCleanPA}
\alias{newCleanPA}
\title{This function cleans up the raw readings from the PA sensors and sets up a data frame.
It assumes that the data has the columns
c('Timestamp', "Names", "PM2.5", "Percent.Difference","Humidity", "Latitude",
"Longitude") in that order. It drops negative PM2.5 values and cleans the time to match LA time.}
\usage{
newCleanPA(data)
}
\arguments{
\item{data}{a .csv of PurpleAir sensor data from the CEHAT website}
}
\value{
a dataframe of the full PurpleAir sensor data, with corrected timezone and PM2.5 levels
}
\description{
This function cleans up the raw readings from the PA sensors and sets up a data frame.
It assumes that the data has the columns
c('Timestamp', "Names", "PM2.5", "Percent.Difference","Humidity", "Latitude",
"Longitude") in that order. It drops negative PM2.5 values and cleans the time to match LA time.
}
