% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{OutCombine}
\alias{OutCombine}
\title{Summarize rows of forward and reverse counts in "out"-object}
\usage{
OutCombine(dataset, samples)
}
\arguments{
\item{dataset}{"out" object from the filtering step of the dada2 analysis}

\item{samples}{vector with sample names}
}
\value{
dataframe with the number of reads per sample before and after filter
}
\description{
The samples argument can be generated with the function \link{CollectData}.
}
