% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilePaths.R
\name{OutputDir}
\alias{OutputDir}
\alias{RepoDir}
\alias{SlurmDir}
\alias{RemoteDir}
\alias{RBScriptDir}
\title{Path to directories in which files are stored locally.}
\usage{
OutputDir()

RepoDir()

SlurmDir()

RemoteDir()

RBScriptDir()
}
\description{
Set the paths in which files should be stored using
\code{options("ntOutDir" = "path/to/analytical/outputs")},
\code{options("ntSlurmDir" = "path/to/slurm/files")}
\code{options("ntRBScriptDir" = "path/to/RevBayes/scripts")}
and
\code{options("ntRepoDir" = "path/where/repos/will/be/checked/out/to")}
}
\details{
Set \code{options("ntRemoteDir" = "path/on/remote/server")} to the directory
on the remote server to which the \code{neotrans} repository is checked out.

These will default respectively to:
\itemize{
\item \code{OutputDir()}: the package's installation location
\item \code{SlurmDir()}: the folder "slurm" in the package's installation location.
\item \code{RepoDir()}: a directory \code{revbayes-repos} placed within the directory
into which the package is installed.
\item \code{RemoteDir()}: \verb{/nobackup/$USER}
}
}
