% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{OutputPlot}
\alias{OutputPlot}
\title{Output a plot to file or graphics device}
\usage{
OutputPlot(figName, width, height, Plot)
}
\arguments{
\item{figName}{Figure filename (without suffix)}

\item{width, height}{Figure dimensions in inches}

\item{Plot}{function to execute, which should produce a plot.}
}
\value{
\code{OutputPlot()} is called for its side-effecits: it launches a new
device of the type specified by \code{getOption("ntOutput")}
(supported: "png", "pdf", "device", where the latter
is the default and plots to the active graphics device).
}
\description{
Output a plot to file or graphics device
}
