% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dataframes.R
\name{overEPA}
\alias{overEPA}
\title{This function takes in a data frame of values from one sensor and returns a dataframe
of the days where the PM2.5 values surpass the EPA threshold.}
\usage{
overEPA(ourData)
}
\arguments{
\item{a}{dataframe of values from one sensor and the name of the PM2.5 category}
}
\value{
a dataframe of the days where the PM2.5 values surpass the EPA threshold.
}
\description{
This function takes in a data frame of values from one sensor and returns a dataframe
of the days where the PM2.5 values surpass the EPA threshold.
}
