% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resources.R
\name{PastRuns}
\alias{PastRuns}
\title{Previous attempts to run an analysis}
\usage{
PastRuns(pID, scriptID)
}
\arguments{
\item{pID}{Character specifying the project identifier to analyse}

\item{scriptID}{Character specifying the script to run}
}
\description{
Previous attempts to run an analysis
}
