% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resources.R
\name{Peek}
\alias{Peek}
\title{View tail of job output}
\usage{
Peek(pID, scriptID, ml = TRUE, lines = 20)
}
\arguments{
\item{pID}{Character specifying the project identifier to analyse}

\item{scriptID}{Character specifying the script to run}

\item{ml}{Logical; if true, calculate the marginal likelihood; if false,
run MCMCMC analysis}
}
\description{
Displays the last \code{lines} lines of the output file produced by a slurm task.
}
