% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-experiments.R
\name{plot_behaviours}
\alias{plot_behaviours}
\title{Plot a behaviour area chart.}
\usage{
plot_behaviours(..., mapping = .gg_check_for_aes(...), facet = TRUE)
}
\arguments{
\item{...}{ggplot format options which can be used to control the plot and
faceting}

\item{mapping}{ggplot \code{aes} mappings which can be used to override the defaults}

\item{facet}{automatically facet the plot base on the experimental design
if you are not happy with the defaults set this to \code{FALSE} and add your own
\code{facet_wrap()} command}
}
\value{
a ggplot with proportion of behaviours exhibited by the population at any
given time during the simulation.
}
\description{
Plot a behaviour area chart.
}
\examples{
with_results_dir(
  here("../scratch/examples/behaviour-comparison"),
  plot_behaviours()
)
}
