% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxp.R
\name{plot_boxp}
\alias{plot_boxp}
\title{Plot the boxplot which is list of the data.}
\usage{
plot_boxp(x, names = NULL)
}
\arguments{
\item{x}{a vector or a list of data.}

\item{names}{a vector of x-axis name.}
}
\description{
Plot the boxplot which is list of the data.
}
