% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotHist.R
\name{plot_hist}
\alias{plot_hist}
\title{Plot a graph of a histogram of an inputted file.}
\usage{
plot_hist(file)
}
\arguments{
\item{file}{an input file.}
}
\description{
Plot a graph of a histogram of an inputted file.
}
