% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare_models.R
\name{plot_model_comparison}
\alias{plot_model_comparison}
\title{Plot and compare the fitted models}
\usage{
plot_model_comparison(task, fitted_models, log_level = NULL)
}
\arguments{
\item{task}{classification task from mlr3}

\item{fitted_models}{fitted models}

\item{log_level}{log level to set. Default is NULL, which means no change in log level. See https://mlr3book.mlr-org.com/logging.html for more details}
}
\value{
ggplot
}
\description{
Plot and compare the fitted models
}
\examples{
{
iris$Species <- as.factor(iris$Species)
models <- c('logistic regression', 'tree', 'gradient boosting', 'svm',
'penalised logistic regression', 'gradient boosting', 'random forest')
fitted_result <- fit_models(iris, 0.5, "Species", models, log_level = 'error')
plot_model_comparison(fitted_result$task, fitted_result$fitted_models)
}
}
