% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-experiments.R
\name{plot_policy}
\alias{plot_policy}
\title{Plot a policy background layer.}
\usage{
plot_policy(
  tables,
  ...,
  mapping = .gg_check_for_aes(...),
  facet = TRUE,
  alpha = FALSE,
  palette = scales::pal_dichromat("BluetoOrange.12")
)
}
\arguments{
\item{tables}{the output of the \code{load_duckdb()} function}

\item{...}{ggplot format options which can be used to control the plot and
faceting}

\item{mapping}{ggplot \code{aes} mappings which can be used to override the defaults}

\item{facet}{automatically facet the plot base on the experimental design
if you are not happy with the defaults set this to \code{FALSE} and add your own
\code{facet_wrap()} command}

\item{alpha}{the default plot is stacked proportions of a particular policy
which would work well as a rug plot underneath a time series. Alternatively
if alpha is \code{TRUE} the policy state is shown as a solid background with
varying hues and degrees of alpha. This is useful for the background of
spaghetti plots of incidence, but this will only be displayed once another
layer is present. In both cases a sensible fill scale needs to be chosen
that works well with discrete data.}

\item{palette}{the palette as a function which takes a number of levels and
returns a colour for that level. In all cases the default (i.e. time zero)
policy will be transparent white so changes will only be obvious for
policies that are a change from the baseline.}
}
\description{
This plots the simulation wide policy in place at any given point in time
in the simulation(s). Multiple simulations may have different policies at
different time (e.g. lockdowns will start and end at different times in
different simulations) These can be represented either by a gradual colour
transition in the background (\code{alpha=TRUE}) or as a stacked bar chart of the
number of simulations in a given policy state (\code{alpha=FALSE})
}
\examples{
tables = load_duckdb("/home/vp22681/Data/ai4ci/0.3.2/behaviour-comparison")
plot_policy(tables)
plot_policy(tables, alpha = TRUE)+ggplot::geom_hline(yintercept=1)
}
