% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotHist.R
\name{PlotHist}
\alias{PlotHist}
\title{Plot a graph of a histogram of an inputted file.}
\usage{
PlotHist(file)
}
\arguments{
\item{file}{an input file.}
}
\description{
Plot a graph of a histogram of an inputted file.
}
