% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predictions.R
\name{PredictWith}
\alias{PredictWith}
\title{Depict predictive power of variable for output}
\usage{
PredictWith(meta, predictor, value, rf, bold = FALSE, ylab = "", col = 1)
}
\arguments{
\item{meta}{Complete metadata available for prediction}

\item{predictor}{Metadata column to use for predicting}

\item{value}{Value to be predicted}

\item{rf}{Random forest, produced using
\verb{randomForest(meta[, <all predictors>], value, importance = TRUE)}.}

\item{ylab, col}{Graphical parameters}
}
\description{
Depict predictive power of variable for output
}
