% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Posterior.R
\name{PriorVsPost}
\alias{PriorVsPost}
\title{Compare prior and posterior distributions}
\usage{
PriorVsPost(pID, scriptID, parameter, priorMean = 0, priorSD = 2)
}
\arguments{
\item{pID}{Character specifying the project identifier to analyse}

\item{scriptID}{Character specifying the script to run}

\item{parameter}{Character string corresponding to column name in
`ParameterFile(pID, scriptID), specifying which parameter to examine.}

\item{priorMean, priorSD}{Mean and SD of a normal prior on \code{parameter}.}
}
\description{
Compare prior and posterior distributions
}
