% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{qq_plot_normal}
\alias{qq_plot_normal}
\title{Generate a qq plot checking for a normal distribution}
\usage{
qq_plot_normal(vec, qs)
}
\arguments{
\item{vec}{A vector containing your actual data}

\item{qs}{A vector containing sample quantiles}
}
\value{
A ggplot object
}
\description{
Generates a plot of sample quantiles to your actual data to check for normality.
It includes a 95% confidence interval.
}
