% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blasttools.R
\name{QSBLAST}
\alias{QSBLAST}
\title{Extract sequence using species name and blast at NCBI}
\usage{
QSBLAST(searchName, seqNumber = 3, blastResults = blastResY)
}
\arguments{
\item{searchName}{string with species name}

\item{seqNumber}{number of sequences from searchName to retain}

\item{blastResults}{dataframe with output from BlastParse function}
}
\description{
Assumes that there is a dataframe that have been created by the
BlastParse function saved in the current workspace. Use QseqGetter
to fetch sequences from this object based on species name and then
use Multiblaster to blast the sequences at NCBI.
}
