% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blasttools.R
\name{QSeqGetter}
\alias{QSeqGetter}
\title{Fetch sequences based on species name and open it}
\usage{
QSeqGetter(speciesName, blastResults = blastResY)
}
\arguments{
\item{speciesName}{string with species name}

\item{blastResults}{dataframe with output from BlastParse function}
}
\value{
an open fasta file with sequence data
}
\description{
This function extracts sequences based on species names and writes to a file named
current_sequence.txt that is opened to be edit with the function
file.edit.
}
\details{
NB! this will overwrite any file named "current_sequence.txt" in
your current working directory.
}
