% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFuncs.R
\name{QueueSim}
\alias{QueueSim}
\title{Queue simulation for remote analysis}
\usage{
QueueSim(
  seed,
  scriptID,
  cores = 16,
  time = getOption("jobTime", (2 * 60 + 25) * 60),
  myMem = 3000,
  myTmp = myMem * 8,
  replace = FALSE,
  withT = TRUE,
  impute = FALSE
)
}
\arguments{
\item{impute}{Logical: should we request imputation simulation, per simImpute.R?}
}
\description{
Queue simulation for remote analysis
}
