% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{ReadTrees}
\alias{ReadTrees}
\title{Read trees from cache}
\usage{
ReadTrees(pID, scriptID)
}
\arguments{
\item{pID}{Character specifying the project identifier to analyse}

\item{scriptID}{Character specifying the script to run}
}
\description{
Read trees from cache
}
