% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RenameSpecies}
\alias{RenameSpecies}
\title{RenameSpecies}
\usage{
RenameSpecies(ref, newName, dataFrame)
}
\arguments{
\item{ref}{a string containing the old name of the species to be renamed.}

\item{newName}{a string containing the new name of the species to be renamed}

\item{dataFrame}{a dataframe containing counts of species with the species names in the first column to the left.}
}
\value{
The original dataframe with the specified species renamed.
}
\description{
Renames a species in a dataframe containing counts of species.
}
