% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilePaths.R
\name{ResultsDir}
\alias{ResultsDir}
\title{Directory containing stored results}
\usage{
ResultsDir(pID, scriptID = NULL)
}
\arguments{
\item{pID}{Character specifying the project identifier to analyse}

\item{scriptID}{Character specifying the script to run}
}
\value{
Character string giving directory path.
}
\description{
Returns the path to where processed results (e.g. \code{.rds} files, summaries)
are written for a given project and script.
Creates the directory if it does not yet exist.
}
\seealso{
Other file-path helpers: 
\code{\link{AnalysisDir}()},
\code{\link{ConvergenceFile}()},
\code{\link{DiskFile}()},
\code{\link{DistanceFile}()},
\code{\link{DummyFile}()},
\code{\link{MarginalFile}()},
\code{\link{MatrixFile}()},
\code{\link{MemFile}()},
\code{\link{PFiles}()},
\code{\link{PPSamplerFile}()},
\code{\link{ParameterFile}()},
\code{\link{ParsEvalFile}()},
\code{\link{ParsimonyFile}()},
\code{\link{ScriptBase}()},
\code{\link{ScriptFile}()},
\code{\link{SlurmFile}()},
\code{\link{SlurmTemplate}()},
\code{\link{StoneFile}()},
\code{\link{StoneOrigin}()},
\code{\link{TimeFile}()},
\code{\link{TreeFiles}()},
\code{\link{TreeSampleFile}()}
}
\concept{file-path helpers}
