% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Posterior.R
\name{RewindRepo}
\alias{RewindRepo}
\title{Restore MCMC files in a git repo to a previous state}
\usage{
RewindRepo(pID, scriptID, lastGoodID)
}
\arguments{
\item{pID}{Character specifying the project identifier to analyse}

\item{scriptID}{Character specifying the script to run}

\item{lastGoodID}{Git identifier denoting (first characters of) last good
commit ID.  A blank entry will return.
"n" will delete relevant files and enqueue a fresh analysis.}
}
\description{
Necessary because some files became corrupted: my suspicion is that
continuing a run on RB<1.4 with RB>1.4 causes an issue when no \code{.heat.ckp}
files are found.
}
\details{
Updates the git repo, deletes cache files, and runs \code{UpdateRecords()}.
}
