% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{runAssayDeltaAnalysis}
\alias{runAssayDeltaAnalysis}
\title{runAssayDeltaAnalysis}
\usage{
runAssayDeltaAnalysis(
  obj,
  assay_name,
  assay_type = "RNA",
  scale,
  patient_id_column = "Patient_ID",
  sample_id_column = "Sample_ID",
  timepoint_column = "Timepoint",
  timepoint_terms = c(baseline = "Baseline", tp2 = "TP2"),
  group_column,
  top_anno,
  top_anno_patient_order,
  palette,
  test_method = c("ttest", "limma"),
  contrasts = "SHORT-LONG",
  pval_cutoff = 0.05,
  run_go = FALSE,
  run_gsea = FALSE,
  logFC_cutoff = 1,
  padj_cutoff = 0.05,
  cancergenes = cancergenes,
  pathwaylists = pathwaylists,
  specialpathwaylists = specialpathwaylists,
  only_sig_genes = F,
  only_show_sig_genes = F,
  save_analysis = T,
  output_dir,
  log = "run assay delta analysis."
)
}
\arguments{
\item{obj}{Analysis. analysis object to be run.}

\item{assay_name}{character. assay name.}

\item{assay_type}{character. assay type, such as "DNA" or "RNA"}

\item{scale}{logic. whether to scale assay.}

\item{patient_id_column}{character. column for patient id.}

\item{sample_id_column}{character. column for sample id.}

\item{timepoint_column}{character. column for timepoint.}

\item{timepoint_terms}{character. timepoint terms pair as delta.}

\item{group_column}{character. column for group.}

\item{top_anno}{HeatmapAnnotation. top annotation, passed to heatmap.}

\item{top_anno_patient_order}{character. patient orders for top annotation.}

\item{palette}{character. colors for KM curve.}

\item{test_method}{character. test method, such as "ttest" or "limma".}

\item{contrasts}{character. specific contrast.}

\item{pval_cutoff}{numeric. p value cutoff for significance.}

\item{run_go}{logic. whether run go analysis.}

\item{run_gsea}{logic. whether run gsea analysis.}

\item{logFC_cutoff}{numeric. log fold change cutoff for signigicant genes.}

\item{padj_cutoff}{numeric. adjusted p value cutoff for significance.}

\item{cancergenes}{character. cancer gene list.}

\item{pathwaylists}{list. pathway list.}

\item{specialpathwaylists}{list. special pathway(geneset) list.}

\item{only_sig_genes}{logic. whether only significant genes will be used.}

\item{only_show_sig_genes}{logic. whether only significant genes will be showed in heatmap.}

\item{save_analysis}{logic. whether to save results locally.}

\item{output_dir}{character. output directory, if missing, use project/analysis_name as directory}

\item{log}{character. any comments.}
}
\value{
Analysis.
}
\description{
run delta analysis about various assay
}
