% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{runAssaySurvivalSignatureAnalysis}
\alias{runAssaySurvivalSignatureAnalysis}
\title{runAssaySurvivalSignatureAnalysis}
\usage{
runAssaySurvivalSignatureAnalysis(
  obj,
  assay_name,
  assay_type = "RNA",
  scale,
  patient_id_column = "Patient_ID",
  sample_id_column = "Sample_ID",
  survival_column,
  status_column,
  survival_type = c("OS"),
  top_anno,
  top_anno_sample_order,
  palette = NULL,
  uni_cox_pvalue_cutoff = 0.01,
  beta_filter = 0.001,
  save_analysis = T,
  output_dir,
  log = "run assay survival signature analysis."
)
}
\arguments{
\item{obj}{Analysis. analysis object to be run.}

\item{assay_name}{character. assay name.}

\item{assay_type}{character. assay type, such as "DNA" or "RNA"}

\item{scale}{logic. whether to scale assay.}

\item{patient_id_column}{character. column for patient id.}

\item{sample_id_column}{character. column for sample id.}

\item{survival_column}{character. column for survival days.}

\item{status_column}{character. column for survival status.}

\item{top_anno}{HeatmapAnnotation. top annotation, passed to heatmap.}

\item{top_anno_sample_order}{character. sample orders for top annotation.}

\item{palette}{character. colors for KM curve.}

\item{uni_cox_pvalue_cutoff}{numeric. p value cutoff for significance of unicox test.}

\item{beta_filter}{numeric. p value cutoff for beta coefficient of unicox test.}

\item{save_analysis}{logic. whether to save results locally.}

\item{output_dir}{character. output directory, if missing, use project/analysis_name as directory}

\item{log}{character. any comments.}

\item{suvival_type}{character. type of survival analysis, such as OS or PFS.}
}
\value{
Analysis.
}
\description{
run assay survival signature analysis
}
