% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{runAssayunSupervisedAnalysis}
\alias{runAssayunSupervisedAnalysis}
\title{runAssayunSupervisedAnalysis}
\usage{
runAssayunSupervisedAnalysis(
  obj,
  assay_name,
  assay_type = "RNA",
  scale = F,
  stablecluster_method = "combine",
  cutFUN = c(ClassDiscovery::cutHclust),
  clusters = NULL,
  nTimes = 100,
  subSampleSize = 0.9,
  subFeatureSize = 0.9,
  patient_id_column = "Patient_ID",
  sample_id_column = "Sample_ID",
  survival_column = NULL,
  status_column = NULL,
  survival_type = c("OS", "PFS"),
  top_anno,
  top_anno_sample_order,
  palette = NULL,
  test_method = c("ttest", "limma"),
  contrasts = NULL,
  pval_cutoff = 0.05,
  run_go = FALSE,
  run_gsea = FALSE,
  logFC_cutoff = 1,
  padj_cutoff = 0.05,
  cancergenes = cancergenes,
  pathwaylists = pathwaylists,
  specialpathwaylists = specialpathwaylists,
  only_sig_genes = F,
  save_analysis = T,
  output_dir,
  log = "run unsupervised assay analysis.",
  ...
)
}
\arguments{
\item{obj}{analysis. an analysis object.}

\item{assay_name}{character. assay name.}

\item{assay_type}{character. type of assay}

\item{scale}{logic. whether to scale assay.}

\item{stablecluster_method}{character. stable cluster method.}

\item{cutFUN}{function. cluster function.}

\item{clusters}{numeric. number of cluters.}

\item{nTimes}{numeric. number of permutation.}

\item{subSampleSize}{numeric. ratio of samples for each permutation.}

\item{subFeatureSize}{numetic. ratio of features for each permutation.}

\item{patient_id_column}{character. column for patient id.}

\item{sample_id_column}{character. column for sample id.}

\item{survival_column}{character. column for survival.}

\item{status_column}{character. column for status.}

\item{survival_type}{character. type of survival analysis.}

\item{top_anno}{HeatmapAnnotation. top annotation, passed to heatmap.}

\item{top_anno_sample_order}{character. patient orders for top annotation.}

\item{palette}{character. colors used for heatmap.}

\item{test_method}{character. test method, such as "ttest" or "limma".}

\item{contrasts}{character. contrast for test.}

\item{pval_cutoff}{numeric. p value cutoff for significance.}

\item{run_go}{logic. whether run go analysis.}

\item{run_gsea}{logic. whether run gsea analysis.}

\item{logFC_cutoff}{numeric. log fold change cutoff for signigicant genes.}

\item{padj_cutoff}{numeric. adjusted p value cutoff for significance.}

\item{cancergenes}{character. cancer gene list.}

\item{pathwaylists}{list. pathway list.}

\item{specialpathwaylists}{list. special pathway(geneset) list.}

\item{only_sig_genes}{logic. whether only significant genes will be used.}

\item{save_analysis}{logic. whether to save results locally.}

\item{output_dir}{character. output directory.}

\item{log}{chracter. any comments.}

\item{...}{list. parameter passed tp difAnalysis}
}
\value{
analysis. contain various contrast result.
}
\description{
run unsupervised assay anlysis
}
