% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{runCCFDifAnalysis}
\alias{runCCFDifAnalysis}
\title{runCCFDifAnalysis}
\usage{
runCCFDifAnalysis(
  obj,
  top_anno,
  top_anno_sample_order,
  palette,
  group,
  block,
  contrasts,
  pval_cutoff = 0.05,
  save_analysis = T,
  output_dir,
  log = "run canopy clone fraction differential analysis."
)
}
\arguments{
\item{obj}{Analysis. analysis object to be run.}

\item{top_anno}{HeatmapAnnotation. top annotation, passed to heatmap.}

\item{top_anno_sample_order}{character. sample orders for top annotation.}

\item{palette}{character. colors for KM curve.}

\item{group}{character. comparison group or column. parameter passed to aov or ttest}

\item{block}{character. comparison block used for paired ttest.}

\item{contrasts}{character. contrasts comparison post aov}

\item{pval_cutoff}{numeric. p value cutoff for significance.}

\item{save_analysis}{logic. whether to save results locally.}

\item{output_dir}{character. output directory, if missing, use project/analysis_name as directory}

\item{log}{character. any comments.}
}
\value{
Analysis.
}
\description{
run differentiate analysis of canopy derived cell fraction
}
