% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{runCINTMBSurvivalAnalysis}
\alias{runCINTMBSurvivalAnalysis}
\title{runCINTMBSurvivalAnalysis
run survival analysis about chromosome instability and tumor mutation burdon}
\usage{
runCINTMBSurvivalAnalysis(
  obj,
  variantstatisticsanalysis,
  CIN_column,
  TMB_column,
  cin_cutoff,
  tmb_cutoff,
  survival_column,
  status_column,
  palette = c(High = "red", Low = "blue"),
  save_analysis = F,
  output_dir,
  log = "run cin and tmb survival analysis."
)
}
\arguments{
\item{obj}{Analysis. analysis object to be run.}

\item{variantstatisticsanalysis}{Analysis. variantstatisticsanalysis object.}

\item{CIN_column}{character. column for chromosome instability.}

\item{TMB_column}{character. column for tumor mutation burdon.}

\item{cin_cutoff}{numeric. chromosome instability cutoff.}

\item{tmb_cutoff}{chromosome instability cutoff.}

\item{survival_column}{character. column for survival days.}

\item{status_column}{character. column for survival status.}

\item{palette}{character. colors for KM curve.}

\item{save_analysis}{logic. whether to save results locally.}

\item{output_dir}{character. output directory, if missing, use project/analysis_name as directory}

\item{log}{character. any comments.}
}
\value{
Analysis.
}
\description{
runCINTMBSurvivalAnalysis
run survival analysis about chromosome instability and tumor mutation burdon
}
