% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{runCloneAnalysis}
\alias{runCloneAnalysis}
\title{runCloneAnalysis
run pyclone Analysis}
\usage{
runCloneAnalysis(
  obj,
  mutect_snps,
  sequenza_snps = NULL,
  sequenza_segments,
  purities,
  overlapping_ref_counts_cutoff = 20,
  overlapping_alt_counts_cutoff = 3,
  nonoverlapping_ref_counts_cutoff = 100,
  nonoverlapping_alt_counts_cutoff = 10,
  gene_locs,
  cancer_genes,
  smg_genes,
  driver_genes,
  pyclone_path = "/home/harryjerry/anaconda3/bin/pyclone-vi",
  save_analysis = F,
  output_dir,
  log = "run sequenza clone analysis."
)
}
\arguments{
\item{obj}{analysis. an analysis object.}

\item{mutect_snps}{data.frame. snps derived from mutect.}

\item{sequenza_snps}{data.frame. snps derived from sequenza.}

\item{sequenza_segments}{data.frame. segments derived from sequenza.}

\item{purities}{data.frame. tumor purity derived from sequenza.}

\item{overlapping_ref_counts_cutoff}{numeric. ref count cutoff for overlapping mutations(Baseline and TP2).}

\item{overlapping_alt_counts_cutoff}{numeric. alt count cutoff for overlapping mutations(Baseline and TP2).}

\item{nonoverlapping_ref_counts_cutoff}{numeric. ref count cutoff for nonoverlapping mutations(Baseline and TP2).}

\item{nonoverlapping_alt_counts_cutoff}{numeric. alt count cutoff for nonoverlapping mutations(Baseline and TP2).}

\item{gene_locs}{data.frame. gene information}

\item{cancer_genes}{character. cancer gene list.}

\item{smg_genes}{character. SMG gene list.}

\item{driver_genes}{character. potential driver genes.}

\item{pyclone_path}{character. path to pyclone}

\item{save_analysis}{logic. whether to save results locally.}

\item{output_dir}{character. output directory.}

\item{log}{character. any comments.}
}
\value{
analysis. contain various contrast result.
}
\description{
runCloneAnalysis
run pyclone Analysis
}
