% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{runCloneunSupervisedAnalysis}
\alias{runCloneunSupervisedAnalysis}
\title{runCloneunSupervisedAnalysis
run unsupervised analysis of Canopy derived clone deconvolution}
\usage{
runCloneunSupervisedAnalysis(
  obj,
  survival_column,
  status_column,
  stablecluster_method = "combine",
  cutFUN = c(ClassDiscovery::cutHclust),
  clusters,
  nTimes = 100,
  subSampleSize = 0.9,
  subFeatureSize = 0.9,
  top_anno,
  top_anno_sample_order,
  palette,
  pval_cutoff = 0.05,
  save_analysis = T,
  output_dir,
  log = "run canopy clone fraction unsuperved analysis."
)
}
\arguments{
\item{obj}{Analysis. analysis object to be run.}

\item{survival_column}{character. column for survival days.}

\item{status_column}{character. column for survival status.}

\item{stablecluster_method}{character. parameter passed to stableCluster}

\item{cutFUN}{function. parameter passed to stableCluster}

\item{clusters}{numeric.parameter passed to stableCluster}

\item{nTimes}{numeric. parameter passed to stableCluster}

\item{subSampleSize}{numeric. parameter passed to stableCluster}

\item{subFeatureSize}{numeric. parameter passed to stableCluster}

\item{top_anno}{HeatmapAnnotation. top annotation, passed to heatmap.}

\item{top_anno_sample_order}{character. sample orders for top annotation.}

\item{palette}{character. colors for KM curve.}

\item{pval_cutoff}{numeric. p value cutoff for significance.}

\item{save_analysis}{logic. whether to save results locally.}

\item{output_dir}{character. output directory, if missing, use project/analysis_name as directory}

\item{log}{character. any comments.}
}
\value{
Analysis.
}
\description{
runCloneunSupervisedAnalysis
run unsupervised analysis of Canopy derived clone deconvolution
}
