% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{runCNVAnalysis}
\alias{runCNVAnalysis}
\title{runCNVAnalysis
run analysis about copy number}
\usage{
runCNVAnalysis(
  obj,
  scale = c("row", "column", "none"),
  method = c("bootstrap", "perturb", "combine"),
  subFeatureSize = 0.8,
  subSampleSize = 1,
  noise = 1,
  cutFUN,
  nTimes = 100,
  clusters = 2,
  verbose = F,
  top_anno,
  top_anno_sample_order,
  sample_order = NULL,
  survival_column = "OS",
  status_column = "Status",
  palette = c("blue", "green", "red"),
  cosmic_tier12_cancer_genes,
  threshold = 2,
  genome = "hg19",
  save_analysis = F,
  output_dir,
  log = "run cnv analysis."
)
}
\arguments{
\item{obj}{Analysis. analysis object to be run.}

\item{scale}{character. parameter passed to stableCluster, see manual of stableCluster.}

\item{method}{character. parameter passed to stableCluster,see manual of stableCluster.}

\item{subFeatureSize}{numeric. parameter passed to stableCluster,see manual of stableCluster.}

\item{subSampleSize}{numeric. parameter passed to stableCluster,see manual of stableCluster.}

\item{noise}{numeric. parameter passed to stableCluster,see manual of stableCluster.}

\item{cutFUN}{function.parameter passed to stableCluster,see manual of stableCluster.}

\item{nTimes}{numeric.parameter passed to stableCluster,see manual of stableCluster.}

\item{clusters}{numeric. parameter passed to stableCluster,see manual of stableCluster.}

\item{verbose}{logic. parameter passed to stableCluster,see manual of stableCluster.}

\item{top_anno}{HeatmapAnnotation. top annotation, passed to heatmap.}

\item{top_anno_sample_order}{character. sample orders for top annotation.}

\item{survival_column}{character. column for survival days.}

\item{status_column}{character. column for survival status.}

\item{palette}{character. colors for KM curve.}

\item{cosmic_tier12_cancer_genes}{data.frame. cosmic tier1 and tier2 mutations, must have column "Gene Symbol".}

\item{threshold}{numeric.}

\item{genome}{character. genome name, such as "hg19", "hg38".}

\item{save_analysis}{logic. whether to save results locally.}

\item{output_dir}{character. output directory, if missing, use project/analysis_name as directory}

\item{log}{character. any comments.}
}
\value{
Analysis.
}
\description{
runCNVAnalysis
run analysis about copy number
}
