% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{runCNVDifAnalysis}
\alias{runCNVDifAnalysis}
\title{runCNVDifAnalysis}
\usage{
runCNVDifAnalysis(
  obj,
  assay_name = "cnv_assay",
  assay_type = "WES",
  level = "cytoband",
  scale,
  patient_id_column = "Patient_ID",
  sample_id_column = "Sample_ID",
  group_column,
  block_column = NULL,
  top_anno,
  top_anno_sample_order,
  palette = NULL,
  test_method = c("ttest", "limma"),
  contrasts = NULL,
  pval_cutoff = 0.05,
  run_go = FALSE,
  run_gsea = FALSE,
  logFC_cutoff = 1,
  padj_cutoff = 0.05,
  cancergenes = cancergenes,
  pathwaylists = pathwaylists,
  specialpathwaylists = specialpathwaylists,
  only_sig_genes = F,
  plot_specific_geneset = F,
  specific_genesets = NULL,
  specific_features = NULL,
  save_analysis = T,
  output_dir,
  log = "run assay differential analysis.",
  ...
)
}
\arguments{
\item{obj}{Analysis. analysis object to be run.}

\item{assay_name}{character. assay name.}

\item{assay_type}{character. assay type, such as "DNA" or "RNA"}

\item{level}{character. what level to call copynumber.}

\item{scale}{logic. whether to scale assay.}

\item{patient_id_column}{character. column for patient id.}

\item{sample_id_column}{character. column for sample id.}

\item{group_column}{character. column for group.}

\item{block_column}{character. column for block.}

\item{top_anno}{HeatmapAnnotation. top annotation, passed to heatmap.}

\item{top_anno_sample_order}{character. patient orders for top annotation.}

\item{palette}{character. colors for KM curve.}

\item{test_method}{character. test method, such as "ttest" or "limma".}

\item{contrasts}{character. specific contrast.}

\item{pval_cutoff}{numeric. p value cutoff for significance.}

\item{run_go}{logic. whether run go analysis.}

\item{run_gsea}{logic. whether run gsea analysis.}

\item{logFC_cutoff}{numeric. log fold change cutoff for signigicant genes.}

\item{padj_cutoff}{numeric. adjusted p value cutoff for significance.}

\item{cancergenes}{character. cancer gene list.}

\item{pathwaylists}{list. pathway list.}

\item{specialpathwaylists}{list. special pathway(geneset) list.}

\item{only_sig_genes}{logic. whether only significant genes will be used.}

\item{plot_specific_geneset}{logic. whether plot heatmap for specific genesets.}

\item{specific_genesets}{character. names of specific geneset, if NULL, significant genesets will be plot.}

\item{specific_features}{character. plot specific features.}

\item{save_analysis}{logic. whether to save results locally.}

\item{output_dir}{character. output directory, if is.null, use project/analysis_name as directory}

\item{log}{character. any comments.}

\item{...}{list. parameter passed to difAnalysis.}
}
\value{
Analysis.
}
\description{
run CNV differential analysis
}
