% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{runFusionAnalysis}
\alias{runFusionAnalysis}
\title{runFusionAnalysis
run analysis of fusions}
\usage{
runFusionAnalysis(
  obj,
  top_fusions = 30L,
  top_anno,
  top_anno_sample_order,
  type_col = NULL,
  sample_order = NULL,
  annotate_frequence = T,
  save_analysis = F,
  output_dir,
  log = "run fusion analysis.",
  ...
)
}
\arguments{
\item{obj}{analysis. an analysis object.}

\item{top_fusions}{numeric. top fusions number.}

\item{top_anno}{HeatmapAnnotation. top annotation, passed to heatmap.}

\item{top_anno_sample_order}{character. patient orders for top annotation.}

\item{type_col}{character. colors for fusion type.}

\item{sample_order}{character.parameter passed to oncoprint.}

\item{annotate_frequence}{logic. whether mark number of fusion detection method.}

\item{save_analysis}{logic. whether to save results locally.}

\item{output_dir}{character. output directory.}

\item{log}{chracter. any comments.}

\item{...}{list. parameter passed oncoprint.}
}
\value{
analysis. contain various contrast result.
}
\description{
runFusionAnalysis
run analysis of fusions
}
