% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{runICGDifAnalysis}
\alias{runICGDifAnalysis}
\title{runICGDifAnalysis
run immune checkpoint gene analysis}
\usage{
runICGDifAnalysis(
  obj,
  icg_info,
  gene_id_column = "Hgnc_Symbol",
  score_method = "zscore",
  survival_column,
  status_column,
  top_anno,
  top_anno_sample_order,
  palette,
  group_column,
  block_column,
  contrasts,
  pval_cutoff = 0.05,
  save_analysis = T,
  output_dir,
  log = "run immune checkpoint differential gene expression analysis."
)
}
\arguments{
\item{obj}{analysis. an analysis object.}

\item{icg_info}{data.frame. immune checkpoint gene information}

\item{gene_id_column}{character. gene ic column from icg_info}

\item{survival_column}{character. suvival days column from patient_info}

\item{status_column}{character. suvival status column from patient_info}

\item{top_anno}{HeatmapAnnotation. top annotation, passed to heatmap.}

\item{top_anno_sample_order}{character. sample orders for top annotation.}

\item{palette}{character. colors used for heatmap.}

\item{group_column}{character. group column from patient_info/sample_info}

\item{block_column}{character. block column from patient_info/sample_info}

\item{contrasts}{character. contrast for test.}

\item{pval_cutoff}{numeric. p value cutoff for significance.}

\item{save_analysis}{logic. whether to save results locally.}

\item{output_dir}{character. output directory.}

\item{log}{chracter. any comments.}
}
\value{
analysis. contain various contrast result.
}
\description{
runICGDifAnalysis
run immune checkpoint gene analysis
}
