% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{runMDLAnalysis}
\alias{runMDLAnalysis}
\title{runMDLAnalysis
run analysis from pyNBS output}
\usage{
runMDLAnalysis(
  obj,
  mdl_info,
  convert2maftools,
  maftools_mutation_conversion_dict,
  mutation_panel_info,
  with_oncokb = T,
  oncokb_cancer_genes_info,
  with_cosmic = T,
  cosmic_mutations,
  include_mdl_only_samples = F,
  include_cnv = F,
  top_anno,
  top_anno_sample_order,
  cancer_genes,
  selected_genes = NULL,
  sample_order = NULL,
  save_analysis = F,
  output_dir,
  log = "run mutational pynbs analysis.",
  ...
)
}
\arguments{
\item{obj}{Analysis. analysis object to be run.}

\item{mdl_info}{data.frame. mdl sample information.}

\item{convert2maftools}{logic. whether convert wes mutation into maftools compatible.}

\item{maftools_mutation_conversion_dict}{character. a named vector mapping cgl mutation and maftools mutation.}

\item{mutation_panel_info}{data.frame. mdl mutation panel information.}

\item{with_oncokb}{logic. whether annotate with oncokb mutation.}

\item{oncokb_cancer_genes_info}{data.frame. oncokb database.}

\item{with_cosmic}{logic. whether annotate with cosmic mutation.}

\item{cosmic_mutations}{data.frame. cosmic mutation database.}

\item{include_mdl_only_samples}{logic. whether include samples with only mdl mutations.}

\item{include_cnv}{logic. whether include cnv in complex_oncoplot.}

\item{top_anno}{HeatmapAnnotation. top annotation, passed to heatmap.}

\item{top_anno_sample_order}{character. sample orders for top annotation.}

\item{cancer_genes}{character. cancer gene list.}

\item{selected_genes}{character. selected genes to show.}

\item{sample_order}{character. sample order for complex onco plot.}

\item{save_analysis}{logic. whether to save results locally.}

\item{output_dir}{character. output directory, if missing, use project/analysis_name as directory}

\item{log}{character. any comments.}

\item{...}{list. parameters passed to complex_onco_plot}
}
\value{
Analysis.
}
\description{
runMDLAnalysis
run analysis from pyNBS output
}
