% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{runMutationalMafAnalysis}
\alias{runMutationalMafAnalysis}
\title{runMutationalMafAnalysis}
\usage{
runMutationalMafAnalysis(
  obj,
  maftools_mutation_conversion_dict = NULL,
  cancer_genes,
  top_anno,
  top_anno_sample_order,
  sample_order = NULL,
  pathways,
  group_column = NULL,
  group_terms = NULL,
  save_analysis = F,
  output_dir,
  log = "run mutational maf analysis.",
  ...
)
}
\arguments{
\item{obj}{Analysis. analysis object to be run.}

\item{maftools_mutation_conversion_dict}{data.frame. dictionary of cgl mutation to maftools compatible mutation.}

\item{cancer_genes}{character. cancer gene list.}

\item{top_anno}{HeatmapAnnotation. top annotation, passed to heatmap.}

\item{top_anno_sample_order}{character. sample orders for top annotation.}

\item{sample_order}{character. oncoplot column order.}

\item{pathways}{list. pathway list containing pathway name and corresponding genes.}

\item{group_column}{character. group column needed compare two diffenrent cohorts}

\item{group_terms}{character. group terms needed for comparisons.}

\item{save_analysis}{logic. whether to save results locally.}

\item{output_dir}{character. output directory, if missing, use project/analysis_name as directory}

\item{log}{character. any comments.}

\item{...}{list. parameters passed to oncoplot.}
}
\value{
Analysis.
}
\description{
run analysis about mutation using maftools
}
