% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{runMutationalSigatureAnalysis}
\alias{runMutationalSigatureAnalysis}
\title{runMutationalSignatureAnalysis}
\usage{
runMutationalSigatureAnalysis(
  obj,
  top_anno,
  top_anno_sample_order,
  extension = 2,
  profile_group,
  rank = 3,
  nrun = 10,
  cutoff = 0.85,
  save_analysis = F,
  output_dir,
  log = "run mutational signature analysis."
)
}
\arguments{
\item{obj}{Analysis. analysis object to be run.}

\item{top_anno}{HeatmapAnnotation. top annotation, passed to heatmap.}

\item{top_anno_sample_order}{character. sample orders for top annotation.}

\item{extension}{numeric.parameter passed mut_matrix.beside mutation site, number of base pair recruited for mutation signature(pattern) analysis.}

\item{profile_group}{character. comparison groups or column. parameter passed to plot_profile_heatmap.}

\item{rank}{numeric. parameter passed to nmf.rank of dimensions.}

\item{nrun}{numeric. parameter passed to nmf.number of runs.}

\item{cutoff}{numeric. parameter passed to rename_nmf_signatures.cutoff of correlation with known signatures.}

\item{save_analysis}{logic. whether to save results locally.}

\item{output_dir}{character. output directory, if missing, use project/analysis_name as directory}

\item{log}{character. any comments.}
}
\value{
Analysis.
}
\description{
run analysis about mutational signature (snp)
}
