% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{runpyNBSAnalysis}
\alias{runpyNBSAnalysis}
\title{runpyNBSAnalysis
run analysis from pyNBS output}
\usage{
runpyNBSAnalysis(
  obj,
  pyNBS_dir,
  survival_column,
  status_column,
  top_anno,
  top_anno_sample_order,
  palette,
  save_analysis = F,
  output_dir,
  log = "run mutational pynbs analysis."
)
}
\arguments{
\item{obj}{Analysis. analysis object to be run.}

\item{pyNBS_dir}{character. path to directory of pyNBS output.}

\item{survival_column}{character. column for survival days.}

\item{status_column}{character. column for survival status.}

\item{top_anno}{HeatmapAnnotation. top annotation, passed to heatmap.}

\item{top_anno_sample_order}{character. sample orders for top annotation.}

\item{palette}{character. colors for KM curve.}

\item{save_analysis}{logic. whether to save results locally.}

\item{output_dir}{character. output directory, if missing, use project/analysis_name as directory}

\item{log}{character. any comments.}
}
\value{
Analysis.
}
\description{
runpyNBSAnalysis
run analysis from pyNBS output
}
