% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{runSurvivalAnalysis}
\alias{runSurvivalAnalysis}
\title{runSurvivalAnalysis
run survival analysis, including univariate/multivariate survival analysis, as well as EIS}
\usage{
runSurvivalAnalysis(
  obj,
  univariate_analysis = T,
  multivariate_analysis = T,
  survival_column,
  status_column,
  covariate_columns,
  extra_covariates,
  palette,
  km_covariates = NULL,
  save_analysis = F,
  output_dir,
  log = "run survival analysis,including univariant/multivariate analysis."
)
}
\arguments{
\item{obj}{Analysis. analysis object to be run.}

\item{univariate_analysis}{logic. run univariate survival analysis.}

\item{multivariate_analysis}{logic. run multivariate survival analysis.}

\item{survival_column}{character. column for survival days.}

\item{status_column}{character. column for survival status.}

\item{covariate_columns}{character. column for covariate passed to univariate,multivariate analysis.}

\item{extra_covariates}{character. column for extra covariate passed to univariate,multivariate analysis.}

\item{palette}{character. colors for KM curve.}

\item{km_covariates}{character. column for additional km plot.}

\item{save_analysis}{logic. whether to save results locally.}

\item{output_dir}{character. output directory, if missing, use project/analysis_name as directory}

\item{log}{character. any comments.}
}
\value{
Analysis.
}
\description{
runSurvivalAnalysis
run survival analysis, including univariate/multivariate survival analysis, as well as EIS
}
