% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{runVariantStatisticsAnalysis}
\alias{runVariantStatisticsAnalysis}
\title{runVariantStatisticsAnalysis
run analysis about varaint statistics (mutation and copynumber)}
\usage{
runVariantStatisticsAnalysis(
  obj,
  cosmic_tier12_cancer_genes,
  genetic_cancer_genes,
  oncokb_cancer_genes,
  save_analysis = F,
  output_dir,
  log = "run variant statistics analysis,including snp, indel and cnv."
)
}
\arguments{
\item{obj}{Analysis. analysis object to be run.}

\item{cosmic_tier12_cancer_genes}{data.frame. cosmic tier1 and tier2 mutations, must have column "Gene Symbol".}

\item{genetic_cancer_genes}{data.frame. cosmic tier1 and tier2 mutations, must have column "gene".}

\item{oncokb_cancer_genes}{data.frame. cosmic tier1 and tier2 mutations, must have column "Hugo Symbol".}

\item{save_analysis}{logic. whether to save results locally.}

\item{output_dir}{character. output directory, if missing, use project/analysis_name as directory}

\item{log}{character. any comments.}
}
\value{
Analysis.
}
\description{
runVariantStatisticsAnalysis
run analysis about varaint statistics (mutation and copynumber)
}
