% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dataframes.R
\name{sensitiveLocations}
\alias{sensitiveLocations}
\title{Sensitive Locations}
\usage{
sensitiveLocations(data, time)
}
\arguments{
\item{the}{input is the data that is outputted in the PAhourly function and a
time a character object of forrmat "yyyy-mm-dd hh:00:00"}
}
\value{
a dataframe that contains sensitive locations and their locations
}
\description{
Sensitive Locations
}
