% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{SeqPlot}
\alias{SeqPlot}
\title{Create a simple plot showing count data from individual samples}
\usage{
SeqPlot(data, sample, title, species)
}
\arguments{
\item{data}{dataframe with count data}

\item{sample}{name of sample to be plotted}

\item{title}{title to be put on the plot}

\item{species}{colname in data that holds species names}
}
\description{
Create a simple plot showing count data from individual samples
}
