% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{set_results_dir}
\alias{set_results_dir}
\title{Set the JPanSim2 output directory.}
\usage{
set_results_dir(dir = NULL)
}
\arguments{
\item{dir}{The directory to set as the JPanSim2 output directory. If missing
a dialog will be opened to select the directory.}
}
\value{
the directory to use.
}
\description{
Set the JPanSim2 output directory.
}
\examples{
set_results_dir(tempdir())
get_results_dir()
}
