% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{set_results_directory}
\alias{set_results_directory}
\title{Set the JPanSim2 output directory.}
\usage{
set_results_directory(dir = NULL)
}
\arguments{
\item{dir}{The directory to set as the JPanSim2 output directory. If missing
a dialog will be opened to select the directory.}
}
\value{
the directory to use.
}
\description{
Set the JPanSim2 output directory.
}
\examples{
set_results_directory(tempdir())
get_results_directory()
}
