% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{set_tables}
\alias{set_tables}
\title{Set the currently loaded simulation output.}
\usage{
set_tables(dir = get_results_dir())
}
\arguments{
\item{dir}{The directory to set as the JPanSim2 output directory. If missing
(and not already set) a dialog will be opened to select the directory.}
}
\description{
Set the currently loaded simulation output.
}
\examples{
set_results_dir(tempdir())
get_tables()
}
