% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilePaths.R
\name{SlurmFile}
\alias{SlurmFile}
\title{Path to generated SLURM script}
\usage{
SlurmFile(pID, scriptID, ml = FALSE)
}
\arguments{
\item{pID}{Character specifying the project identifier to analyse}

\item{scriptID}{Character specifying the script to run}

\item{ml}{Logical; whether this corresponds to a marginal-likelihood run.}
}
\value{
Character string giving file path.
}
\description{
Constructs the path to the SLURM job script for a given project and script.
}
\seealso{
Other file-path helpers:
\code{\link[=AnalysisDir]{AnalysisDir()}},
\code{\link[=ConvergenceFile]{ConvergenceFile()}},
\code{\link[=DiskFile]{DiskFile()}},
\code{\link[=DistanceFile]{DistanceFile()}},
\code{\link[=DummyFile]{DummyFile()}},
\code{\link[=MarginalFile]{MarginalFile()}},
\code{\link[=MatrixFile]{MatrixFile()}},
\code{\link[=MemFile]{MemFile()}},
\code{\link[=PFiles]{PFiles()}},
\code{\link[=PPSamplerFile]{PPSamplerFile()}},
\code{\link[=ParameterFile]{ParameterFile()}},
\code{\link[=ParsEvalFile]{ParsEvalFile()}},
\code{\link[=ParsimonyFile]{ParsimonyFile()}},
\code{\link[=ResultsDir]{ResultsDir()}},
\code{\link[=ScriptBase]{ScriptBase()}},
\code{\link[=ScriptFile]{ScriptFile()}},
\code{\link[=SlurmTemplate]{SlurmTemplate()}},
\code{\link[=StoneFile]{StoneFile()}},
\code{\link[=StoneOrigin]{StoneOrigin()}},
\code{\link[=TimeFile]{TimeFile()}},
\code{\link[=TreeFiles]{TreeFiles()}},
\code{\link[=TreeSampleFile]{TreeSampleFile()}}
}
\concept{file-path helpers}
