% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resources.R
\name{SlurmLog}
\alias{SlurmLog}
\title{Read log of completed SLURM jobs}
\usage{
SlurmLog()
}
\value{
\code{SlurmLog()} returns a data.frame with details of resource allocation
to previous slurm jobs.
}
\description{
Read log of completed SLURM jobs
}
