% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resources.R
\name{SlurmQueue}
\alias{SlurmQueue}
\title{Read current slurm queue on remote host}
\usage{
SlurmQueue(session = SshSession())
}
\value{
\code{SlurmQueue()} returns a data.frame containing details of queued
jobs.
}
\description{
Read current slurm queue on remote host
}
