% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilePaths.R
\name{SlurmTemplate}
\alias{SlurmTemplate}
\title{Path to SLURM submission template}
\usage{
SlurmTemplate(ml, mpi = TRUE)
}
\arguments{
\item{ml}{Logical; whether this is a marginal-likelihood run.}

\item{mpi}{Logical; whether the analysis uses MPI parallelization.}
}
\value{
Character string giving path to SLURM template.
}
\description{
Returns the template script appropriate for a given analysis mode.
}
\seealso{
Other file-path helpers:
\code{\link[=AnalysisDir]{AnalysisDir()}},
\code{\link[=ConvergenceFile]{ConvergenceFile()}},
\code{\link[=DiskFile]{DiskFile()}},
\code{\link[=DistanceFile]{DistanceFile()}},
\code{\link[=DummyFile]{DummyFile()}},
\code{\link[=MarginalFile]{MarginalFile()}},
\code{\link[=MatrixFile]{MatrixFile()}},
\code{\link[=MemFile]{MemFile()}},
\code{\link[=PFiles]{PFiles()}},
\code{\link[=PPSamplerFile]{PPSamplerFile()}},
\code{\link[=ParameterFile]{ParameterFile()}},
\code{\link[=ParsEvalFile]{ParsEvalFile()}},
\code{\link[=ParsimonyFile]{ParsimonyFile()}},
\code{\link[=ResultsDir]{ResultsDir()}},
\code{\link[=ScriptBase]{ScriptBase()}},
\code{\link[=ScriptFile]{ScriptFile()}},
\code{\link[=SlurmFile]{SlurmFile()}},
\code{\link[=StoneFile]{StoneFile()}},
\code{\link[=StoneOrigin]{StoneOrigin()}},
\code{\link[=TimeFile]{TimeFile()}},
\code{\link[=TreeFiles]{TreeFiles()}},
\code{\link[=TreeSampleFile]{TreeSampleFile()}}
}
\concept{file-path helpers}
