% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{SpindlePlot}
\alias{SpindlePlot}
\title{Spindle plot}
\usage{
SpindlePlot(
  x,
  nBin = 20,
  width = 1,
  xlab = "",
  log = FALSE,
  Behind = function() {
 },
  clip,
  weights = NULL,
  ...
)
}
\arguments{
\item{x}{data.frame or matrix in which columns correspond to distinct values
to plot.}

\item{nBin}{Integer giving number of bins}

\item{width}{Numeric giving relative width of each spindle}

\item{log}{Logical specifying whether to log-transform the y axis}

\item{Behind}{Function to be executed before drawing the spindles}

\item{weights}{Numeric specifying weight of each observation, for a
weighted mean.}

\item{\dots}{Additional arguments to \code{plot()}}

\item{cols}{Vector giving colour for each spindle}
}
\description{
\code{SpindlePlot()} creates a histogram arranged analogously to a violin plot
}
