% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_urn}
\alias{split_urn}
\title{Splits a vector of JPanSim2 URNs into a tibble of key-value pairs.}
\usage{
split_urn(urns)
}
\arguments{
\item{urns}{A character vector of URNs, where each URN is expected to have
components separated by colons.}
}
\value{
A tibble where each column corresponds to a key from the URNs, and
rows represent the associated values. If a URN has fewer than two components,
it is ignored.
}
\description{
This function processes URNs in the format key:value:key:value and converts
them into a tibble where each row corresponds to a URN and columns represent
the extracted keys and their associated values.
}
\examples{
urns <- c("key1:value1:key2:value2", "keyA:valueA:keyB:valueB")
split_urn(urns)

}
\keyword{internal}
