% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resources.R
\name{SQ}
\alias{SQ}
\title{squeue
Output the current slurm queue on the remote host}
\usage{
SQ(k = NULL)
}
\arguments{
\item{k}{Sort criterion for job list}
}
\description{
squeue
Output the current slurm queue on the remote host
}
