% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSH.R
\name{SshSession}
\alias{SshSession}
\title{Start ssh session with remote server}
\usage{
SshSession()
}
\value{
\code{SshSession()} is called for its side-effect of beginning a new
ssh session and registering this in the package environment for future
calls to the remote.
}
\description{
Start ssh session with remote server
}
