% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dataframes.R
\name{summarySG}
\alias{summarySG}
\title{Hourly Summary Data for South Gate}
\usage{
summarySG(data)
}
\arguments{
\item{data}{the output dataframe of the hourlyPA function}
}
\value{
a dataframe of the South Gate data, including \emph{weighted} sensor averages, sensor median, the max, and the min, and active count.
}
\description{
This function aggregates HOURLYPA data (output from the hourlyPA function) by hour for mean, median, and count of sensors
It assumes that the data has the columns 'timestamp', 'day', 'hour' 'PM2.5.
}
