% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2analysis.R
\name{SumRes}
\alias{SumRes}
\title{Create dataframe with sequences, counts and species}
\usage{
SumRes(blastRes, counts, taxGroup)
}
\arguments{
\item{blastRes}{dataframe created with the blastParse function}

\item{counts}{dataframe with count data}

\item{taxGroup}{name for taxonomic group to be extracted from the
total data}
}
\value{
resCount dataframe with sequence, taxonomy and counts
}
\description{
This functions merge annotation of sequences with count matrix and
facilitates the creation of data sets of specified species groups.
}
\details{
NB! There might be a need to modify the function to create new
filter options as certain species groups can not be extracted
by a simple selection at class level. Check the fish implementation
in the code of the function.
}
