% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aesthetic.R
\name{TaxonCol}
\alias{TaxonCol}
\title{Taxon colour
Return an appropriate colour for each higher taxon}
\usage{
TaxonCol(x = 1:4)
}
\description{
Taxon colour
Return an appropriate colour for each higher taxon
}
