% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_summary.R
\name{tb_summary}
\alias{tb_summary}
\title{Plot the sammary table from a list of data.}
\usage{
tb_summary(x)
}
\arguments{
\item{x}{a list of data.n}
}
\description{
Plot the sammary table from a list of data.
}
