% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{TranslateTaxa}
\alias{TranslateTaxa}
\title{TranslateTaxa}
\usage{
TranslateTaxa(nameVector)
}
\arguments{
\item{nameVector}{a vector of scientific latin names.}
}
\value{
a vector that contains Swedish vernacular names corresponding to the input vector of scientific latin name
}
\description{
Takes a vector of latin names of any length and returns a vector containing Swedish vernacular names.
}
