% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilePaths.R
\name{TreeFiles}
\alias{TreeFiles}
\title{List tree-file paths}
\usage{
TreeFiles(pID, scriptID, compressed = FALSE)
}
\arguments{
\item{pID}{Character specifying the project identifier to analyse}

\item{scriptID}{Character specifying the script to run}

\item{compressed}{Logical; whether to list compressed files.}
}
\value{
Character vector of tree-file paths.
}
\description{
Returns a vector of paths to \code{.trees} (or compressed \code{.tar.gz}) files
generated by an analysis.
}
\seealso{
Other file-path helpers:
\code{\link[=AnalysisDir]{AnalysisDir()}},
\code{\link[=ConvergenceFile]{ConvergenceFile()}},
\code{\link[=DiskFile]{DiskFile()}},
\code{\link[=DistanceFile]{DistanceFile()}},
\code{\link[=DummyFile]{DummyFile()}},
\code{\link[=MarginalFile]{MarginalFile()}},
\code{\link[=MatrixFile]{MatrixFile()}},
\code{\link[=MemFile]{MemFile()}},
\code{\link[=PFiles]{PFiles()}},
\code{\link[=PPSamplerFile]{PPSamplerFile()}},
\code{\link[=ParameterFile]{ParameterFile()}},
\code{\link[=ParsEvalFile]{ParsEvalFile()}},
\code{\link[=ParsimonyFile]{ParsimonyFile()}},
\code{\link[=ResultsDir]{ResultsDir()}},
\code{\link[=ScriptBase]{ScriptBase()}},
\code{\link[=ScriptFile]{ScriptFile()}},
\code{\link[=SlurmFile]{SlurmFile()}},
\code{\link[=SlurmTemplate]{SlurmTemplate()}},
\code{\link[=StoneFile]{StoneFile()}},
\code{\link[=StoneOrigin]{StoneOrigin()}},
\code{\link[=TimeFile]{TimeFile()}},
\code{\link[=TreeSampleFile]{TreeSampleFile()}}
}
\concept{file-path helpers}
