% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parsimony.R
\name{TreeSearch}
\alias{TreeSearch}
\title{Queue parsimony search on Slurm remote}
\usage{
TreeSearch(pID, replace = FALSE)
}
\arguments{
\item{pID}{Character specifying the project identifier to analyse}

\item{replace}{Logical: should a corresponding job be cancelled in order to
run this job instead?}
}
\description{
Queue parsimony search on Slurm remote
}
\seealso{
\code{\link[=MakeSlurm]{MakeSlurm()}}
}
