% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{TreeSimBox}
\alias{TreeSimBox}
\alias{TreeSimPlot}
\title{Box plot of tree similarity}
\usage{
TreeSimBox(pID, models, output = NULL, sampleSize = 256, method = "quartet")

TreeSimPlot(pID, models, pdf = FALSE, sampleSize = 128)
}
\arguments{
\item{pID}{Character specifying the project identifier to analyse}
}
\description{
Box plot of tree similarity
}
