% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeComparison.R
\name{TreeSimilarities}
\alias{TreeSimilarities}
\title{Compare trees with well corroborated tree}
\usage{
TreeSimilarities(
  pID,
  models,
  method = "quartet",
  sampleSize = 256,
  useCache = TRUE
)
}
\arguments{
\item{useCache}{Logical sppecifing whether to use cached results. Set to
\code{FALSE} to regenerate the cache, e.g. if new trees are available.}
}
\description{
Compare trees with well corroborated tree
}
