% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{TreeSimSpindle}
\alias{TreeSimSpindle}
\title{Spindle plot of tree similarity}
\usage{
TreeSimSpindle(
  pID,
  models,
  output = NULL,
  sampleSize = 256,
  method = "quartet",
  nBin = 12,
  width = 1,
  clip,
  topMar = 1
)
}
\arguments{
\item{pID}{Character specifying the project identifier to analyse}

\item{nBin}{Integer giving number of bins}

\item{width}{Numeric giving relative width of each spindle}
}
\description{
Spindle plot of tree similarity
}
