% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{TreeSimSpindleRow}
\alias{TreeSimSpindleRow}
\title{Plot a row of tree similarity spindles}
\usage{
TreeSimSpindleRow(
  pID,
  marginals,
  models,
  silID,
  Annotate1 = function() {
 }
)
}
\arguments{
\item{pID}{Character specifying the project identifier to analyse}

\item{models}{Which script IDs should be included in the plot?}

\item{silID}{Phylopic uuid identifier for silhouette to add}

\item{Annotate1}{Function to be called after first panel is drawn,
perhaps to add annotation or panel label}
}
\description{
A useful figure component
}
