% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{TrimCutoffs}
\alias{TrimCutoffs}
\title{Remove sequences and counts based on cutoffs}
\usage{
TrimCutoffs(countsObject, total = 0, within = 0)
}
\arguments{
\item{countsObject}{dataframe with columns 1:5 with sequence info (unique sequence references, notes, Swedish name, Latin name and percentage out of all sequences in the data frame) and the rest of the columns of samples with counts of the sequences.}

\item{total}{percentage cutoff specifying the minimum percentage of counts a sequence has to be represented by of out of the total number of counts in the data frame. Sequences that fall below the total\% get removed.}

\item{within}{percentage cutoff specifying the minimum percentage of counts a sequence has to be represented by of out of the total number of counts in a sample (a column in the data fram). Sequences that fall below the within\% get gets reduced to 0 within that specific sample.}
}
\value{
dataframe with columns 1:5 with sequence info (sequence reference, notes, Swedish name, Latin name and percentage out of all sequences in the data frame) and the rest of the columns of samples with counts of the sequences.
}
\description{
Remove sequences and counts based on cutoffs
}
