% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{try_mutate}
\alias{try_mutate}
\title{Attempts to apply a mutation to a data frame, handling errors gracefully.}
\usage{
try_mutate(df, ...)
}
\arguments{
\item{df}{A data frame to which the mutation will be applied.}

\item{...}{Additional arguments passed to \code{mutate()}, specifying the mutation(s).}
}
\value{
The mutated data frame if successful; the original data frame if an error occurs.
If the result is neither a data frame nor an error, the raw output is returned.
}
\description{
Attempts to apply a mutation to a data frame, handling errors gracefully.
}
\examples{
df <- tibble::tibble(x = 1:5)
try_mutate(df, y = x * 2) # Successfully adds a new
try_mutate(df, w = stop()) # Error occurs, returns original df
try_mutate(df, v = x + "a") # Error occurs, returns original df

}
\keyword{internal}
