% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttests.R
\name{ttests}
\alias{ttests}
\title{This function takes in two arrays of data (with corresponding
days) and the city name (the non-south gate city name) and
returns the results after conducting t-tests on them}
\usage{
ttests(otherCityData, SGdata, cityName)
}
\arguments{
\item{two}{arrays of data with matching days and the name of the non-South Gate city}
}
\value{
it returns the t test results of the data sets
}
\description{
This function takes in two arrays of data (with corresponding
days) and the city name (the non-south gate city name) and
returns the results after conducting t-tests on them
}
