% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckComplete.R
\name{UpdateRecords}
\alias{UpdateRecords}
\title{Fetch ML and MCMCMC results from GitHub.}
\usage{
UpdateRecords(
  pID,
  scriptID,
  searchRemote = FALSE,
  forgetCache = FALSE,
  makeSlurm = FALSE
)
}
\arguments{
\item{searchRemote}{Logical: if \verb{git fetch} does not retrieve new results,
should we check for uncommitted changes on the remote host?}

\item{forgetCache}{Logical: if no new results are found, should we
recalculate cached statistics anyway?}
}
\description{
If new MCMCMC result are available, compute and cache convergence statistics.
}
\details{
Parameters and trees are written to cache using \code{UpdateResults()}, called
from within this function.
}
