% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{urn_groups}
\alias{urn_groups}
\title{Extracts unique group identifiers from a vector of JPanSim2 URNs.}
\usage{
urn_groups(urns)
}
\arguments{
\item{urns}{A character vector of URNs, where each URN is expected
to have components separated by colons.}
}
\value{
A character vector of unique group identifiers extracted from the URNs.
}
\description{
URNs in JPanSim2 will be in the form key:value:key:value:key:value.
This function retrieves all the keys.
}
\examples{
urns <- c("key1:value1:key2:value2", "keyA:valueA:keyB:valueB")
urn_groups(urns)

}
\keyword{internal}
