% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{visual_normality_check}
\alias{visual_normality_check}
\title{Build a comparison & plot to see visual normality check}
\usage{
visual_normality_check(vec)
}
\arguments{
\item{vec}{A vector containing your actual data}
}
\value{
A ggplot object
}
\description{
Constructs sample quantiles & compares them to your data to enable a quick visual check for normality
}
