% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{with_results_dir}
\alias{with_results_dir}
\title{Temporarily set results directory and evaluate an expression}
\usage{
with_results_dir(dir, expr)
}
\arguments{
\item{dir}{the results directory}

\item{expr}{the expression to evaluate}
}
\value{
the result of the expression
}
\description{
Temporarily set results directory and evaluate an expression
}
\examples{
with_results_dir(tempdir(), cat(get_results_dir()))
}
