% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilePaths.R
\name{AnalysisDir}
\alias{AnalysisDir}
\title{Directory containing analysis output}
\usage{
AnalysisDir(pID, scriptID = NULL)
}
\arguments{
\item{pID}{Character specifying the project identifier to analyse}

\item{scriptID}{Character specifying the script to run}
}
\value{
Character string giving directory path.
}
\description{
Returns the path to the directory containing RevBayes log and tree files
for a specific project and script.
}
\seealso{
Other file-path helpers: 
\code{\link{ConvergenceFile}()},
\code{\link{DiskFile}()},
\code{\link{DistanceFile}()},
\code{\link{DummyFile}()},
\code{\link{MarginalFile}()},
\code{\link{MatrixFile}()},
\code{\link{MemFile}()},
\code{\link{PFiles}()},
\code{\link{PPSamplerFile}()},
\code{\link{ParameterFile}()},
\code{\link{ParsEvalFile}()},
\code{\link{ParsimonyFile}()},
\code{\link{ResultsDir}()},
\code{\link{ScriptBase}()},
\code{\link{ScriptFile}()},
\code{\link{SlurmFile}()},
\code{\link{SlurmTemplate}()},
\code{\link{StoneFile}()},
\code{\link{StoneOrigin}()},
\code{\link{TimeFile}()},
\code{\link{TreeFiles}()},
\code{\link{TreeSampleFile}()}
}
\concept{file-path helpers}
