% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2analysis.R
\name{BlastParse}
\alias{BlastParse}
\title{Parse blast output}
\usage{
BlastParse(dgeList, blastRes)
}
\arguments{
\item{dgeList}{count data in the form of DGEList}

\item{blastRes}{file with blast results assumes blastoutput option
-outfmt "6 qseqid sseqid pident length mismatch gapopen qstart qend
sstart send evalue bitscore staxids sscinames scomnames"}
}
\value{
seqTax a dataframe with annotation of results from sequence
comparisons using blast.
}
\description{
This function will create a dataframe with sequence names,
sequences similarity from blast analysis together with taxonomic
information from NCBI.
}
\details{
NB! for optimal performance make sure to get a token from NCBI as
the function relies on a online query to the NCBI taxonomic
database. It hence also needs an active internet connection to
work.
}
\examples{
fastqR1 <- system.file("extdata", "exampleFq_R1.fastq.gz",
package = "MetaBAnalysis")
fastqR2 <- system.file("extdata", "exampleFq_R2.fastq.gz",
package = "MetaBAnalysis")
parseTest <- DadaAnalysis(fastqR1, fastqR2, muThread = FALSE)
dfForward <- as.data.frame(t(parseTest))
yForward <- edgeR::DGEList(dfForward)
BlastParse(dgeList = yForward, blastRes = system.file("extdata", "test.out", package = "MetaBAnalysis"))
}
