% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resources.R
\name{Collect}
\alias{Collect}
\title{Process analyses that have completed on remote server}
\usage{
Collect(days = 1)
}
\arguments{
\item{days}{Numeric specifying how many days since last collection to
retrieve}
}
\value{
\code{Collect()} is called for its side effects of pulling remote results
from the server, and updating the local cache accordingly.
}
\description{
Process analyses that have completed on remote server
}
