% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CollectData}
\alias{CollectData}
\title{Collect sample names and paths}
\usage{
CollectData(directory = "../Filtered_data")
}
\arguments{
\item{directory}{path to directory with filtered data}
}
\value{
list with sample paths and sample names
}
\description{
Collect sample names and paths
}
