% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CombineSpecies}
\alias{CombineSpecies}
\title{CombineSpecies}
\usage{
CombineSpecies(species1, species2, dataFrame)
}
\arguments{
\item{species1}{a string containing the name of the first species to be combined (this name will be kept).}

\item{species2}{a string containing the name of the second species to be combined.}

\item{dataFrame}{a dataframe containing counts of species with the species names in the first column to the left.}
}
\value{
the given dataframe with the two species counts combined under the first name given.
}
\description{
Takes the names of two species and a dataframe of counts and returns the dataframe with the counts of the two species combined.
}
