% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dataframes.R
\name{compareSensors}
\alias{compareSensors}
\title{Above/Below Median}
\usage{
compareSensors(data, type = c("a", "b"))
}
\arguments{
\item{data}{output of the hourlyPA function (a dataframe of hourly PurpleAir sensor data)}

\item{type}{a single-character string indicating whether values above or below the median should be returned}
}
\value{
a dataframe that details how many readings are above/below the median for each sensor, along with that sensors total readings
}
\description{
This reports which sensors were down and on which days that they went
}
