% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilePaths.R
\name{ConvergenceFile}
\alias{ConvergenceFile}
\title{Path to convergence-diagnostic file}
\usage{
ConvergenceFile(pID, scriptID)
}
\arguments{
\item{pID}{Character specifying the project identifier to analyse}

\item{scriptID}{Character specifying the script to run}
}
\value{
Character string giving \code{.txt} path summarizing convergence checks.
}
\description{
Path to convergence-diagnostic file
}
\seealso{
Other file-path helpers: 
\code{\link{AnalysisDir}()},
\code{\link{DiskFile}()},
\code{\link{DistanceFile}()},
\code{\link{DummyFile}()},
\code{\link{MarginalFile}()},
\code{\link{MatrixFile}()},
\code{\link{MemFile}()},
\code{\link{PFiles}()},
\code{\link{PPSamplerFile}()},
\code{\link{ParameterFile}()},
\code{\link{ParsEvalFile}()},
\code{\link{ParsimonyFile}()},
\code{\link{ResultsDir}()},
\code{\link{ScriptBase}()},
\code{\link{ScriptFile}()},
\code{\link{SlurmFile}()},
\code{\link{SlurmTemplate}()},
\code{\link{StoneFile}()},
\code{\link{StoneOrigin}()},
\code{\link{TimeFile}()},
\code{\link{TreeFiles}()},
\code{\link{TreeSampleFile}()}
}
\concept{file-path helpers}
