% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2analysis.R
\name{DadaAnalysis}
\alias{DadaAnalysis}
\title{Convenience function to run dada2}
\usage{
DadaAnalysis(
  primerData,
  muThread = TRUE,
  justConcatenate = FALSE,
  minOverlap = 5
)
}
\arguments{
\item{primerData}{list containing information about samples to be run, can be generated with the function \link{CollectData}. The samples in the resulting object will be named after primerData$Samples}

\item{muThread}{Should the analysis use multithreads for analysis}

\item{justConcatenate}{if reads pairs do not overlap set to TRUE}

\item{minOverlap}{minimum overlap between reads when merging pairs}
}
\value{
a matrix with count for all inferred sequence variants
}
\description{
This largely runs several of the standard steps in a dada2 analysis
with default parameters.
}
\examples{
filteredDataFolder <- system.file("extdata", package = "MetAlPipe")
exampleCollectData <- CollectData(filteredDataFolder)
DadaAnalysis(exampleCollectData, muThread = FALSE)

}
