% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2analysis.R
\name{DadaAnalysis}
\alias{DadaAnalysis}
\title{Convience function to run dada2}
\usage{
DadaAnalysis(forward, reverse, muThread = TRUE, justConcatenate = FALSE)
}
\arguments{
\item{forward}{vector with file names for filtered forward reads}

\item{reverse}{vector with file names for filtered reverse reads}

\item{muthread}{Should the analysis use multithreads for analysis}
}
\value{
a matrix with count for all inferred sequence variants
}
\description{
This largely run several of the standard steps in a dada2 analysis
with default parameters.
}
\examples{
fastqR1 <- system.file("extdata", "exampleFq_R1.fastq.gz",
package = "MetaBAnalysis")
fastqR2 <- system.file("extdata", "exampleFq_R2.fastq.gz",
package = "MetaBAnalysis")
DadaAnalysis(fastqR1, fastqR2, muthread = FALSE)
}
