\name{distance}
\alias{distance}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Distance between two Locations}
\description{
This function takes in four parameters: lat_1,long_1,lat_2,long_2.
It then calculates the distance between these two locations using
the Haversine formula.
}
\usage{
distance(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
the latitude and longitude of one location and the latitude
and longitude of another location
}
}
\value{the distance between two points in space in miles}

