% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilePaths.R
\name{DistanceFile}
\alias{DistanceFile}
\title{Path to tree-distance results}
\usage{
DistanceFile(pID, model1, model2)
}
\arguments{
\item{pID}{Character specifying the project identifier to analyse}

\item{model1, model2}{Names of two models to compare.}
}
\value{
Character string giving \code{.rds} path containing distance metrics.
}
\description{
Path to tree-distance results
}
\seealso{
Other file-path helpers: 
\code{\link{AnalysisDir}()},
\code{\link{ConvergenceFile}()},
\code{\link{DiskFile}()},
\code{\link{DummyFile}()},
\code{\link{MarginalFile}()},
\code{\link{MatrixFile}()},
\code{\link{MemFile}()},
\code{\link{PFiles}()},
\code{\link{PPSamplerFile}()},
\code{\link{ParameterFile}()},
\code{\link{ParsEvalFile}()},
\code{\link{ParsimonyFile}()},
\code{\link{ResultsDir}()},
\code{\link{ScriptBase}()},
\code{\link{ScriptFile}()},
\code{\link{SlurmFile}()},
\code{\link{SlurmTemplate}()},
\code{\link{StoneFile}()},
\code{\link{StoneOrigin}()},
\code{\link{TimeFile}()},
\code{\link{TreeFiles}()},
\code{\link{TreeSampleFile}()}
}
\concept{file-path helpers}
