% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DropSpecies}
\alias{DropSpecies}
\title{DropSpecies}
\usage{
DropSpecies(name, dataFrame)
}
\arguments{
\item{dataFrame}{a dataframe containing counts of species with the species names in the first column to the left.}

\item{species}{a string containing the name of the species to be removed.}
}
\value{
the given dataframe without the species specified when calling the function.
}
\description{
Takes the name of a species and a dataframe of counts and returns the dataframe without the given species.
}
