% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ExportFasta}
\alias{ExportFasta}
\title{Export DNA sequence as fasta file}
\usage{
ExportFasta(countData, fileName, minLength = 50, maxLength = 1000)
}
\arguments{
\item{countData}{a dataframe with DNA sequences as row names}

\item{fileName}{the name of the file to write data to}

\item{minLength}{a minimum length criteria for sequences to pass through}

\item{maxLength}{a maximum length criteria for sequences to pass through}
}
\value{
the number of sequences written to the file as well as the actual file on disk
}
\description{
Export DNA sequence as fasta file
}
\examples{
countData <- data.frame(A = 1:3, B = 4:6, C = 7:9, row.names = c("ACTG", "TTAG", "GGCA"))
ExportFasta(countData, fileName = "junk.fa")
}
