% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{FiltTrimWrap}
\alias{FiltTrimWrap}
\title{Convenience function to run dada2}
\usage{
FiltTrimWrap(primerData)
}
\arguments{
\item{primerData}{list containing information about samples to be run, can be generated with the function \link{CollectData}.}
}
\value{
a matrix with counts of trimmed and untrimmed sequences for every sample
}
\description{
This largely runs several of the standard steps in a dada2 analysis
with default parameters.
}
\examples{
fastqR1 <- system.file("extdata", "exampleFq_R1.fastq.gz",
package = "MetaBAnalysis")
fastqR2 <- system.file("extdata", "exampleFq_R2.fastq.gz",
package = "MetaBAnalysis")
DadaAnalysis(fastqR1, fastqR2, muThread = FALSE)

}
