% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dataframes.R
\name{hourlyPA}
\alias{hourlyPA}
\title{Aggregate Data by Hour}
\usage{
hourlyPA(data, new)
}
\arguments{
\item{data}{a CLEANED dataframe of PurpleAir sensor data a boolean specifying if we are using the new}

\item{new}{a boolean specifying if we are using the new data (in which case you put TRUE) or the old data (in which case you put FALSE)}
}
\value{
a dataframe of the hourly PA sensor data, with sensor names, days of the week, hour, and time of day
}
\description{
This function aggregates the frequent readings from each sensor and averages them by hour.
It assumes that the data has the columns 'timestamp' and 'longitude'.
It also uses an EPA correction factor for PurpleAir sensors.
Any readings with faulty channel A/B values will be dropped
}
