% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqiCalculation.R
\name{indexCalculation}
\alias{indexCalculation}
\title{This function takes in five parameters: highIndexBreakpoint,lowIndexBreakpoint,
highConcentrationBreakpoint,lowConcentrationBreakpoint,pm25Concentration, in
that order. It then calculates the AQI guven the specified indices.}
\usage{
indexCalculation(
  highIndexBreakpoint,
  lowIndexBreakpoint,
  highConcentrationBreakpoint,
  lowConcentrationBreakpoint,
  pm25Concentration
)
}
\arguments{
\item{highIndexBreakpoint, lowIndexBreakpoint, }{highConcentrationBreakpoint,lowConcentrationBreakpoint,pm25Concentration}
}
\value{
it returns the AQI that corresponds to the given PM2.5 measurement
}
\description{
This function takes in five parameters: highIndexBreakpoint,lowIndexBreakpoint,
highConcentrationBreakpoint,lowConcentrationBreakpoint,pm25Concentration, in
that order. It then calculates the AQI guven the specified indices.
}
