% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Posterior.R
\name{InformationGain}
\alias{InformationGain}
\title{Increase in information density in prior vs posterior.}
\usage{
InformationGain(
  projects = KiProjects(),
  scripts,
  parameter,
  metaEntry,
  priorMean = 0,
  priorSD = 2
)
}
\arguments{
\item{projects}{Which projects to use}

\item{parameter}{Character string corresponding to column name in
`ParameterFile(pID, scriptID), specifying which parameter to examine.}

\item{metaEntry}{Metadata item to regress with \code{parameter}}

\item{priorMean, priorSD}{Mean and SD of a normal prior on \code{parameter}.}
}
\value{
\code{InformationGain()} returns the spearman's correlation of
\code{metaEntry} with \code{parameter}

\code{InformationGain()} returns the results of a Spearman correlation
conducted with \code{cor.test(method = "spearman")} between the prior/posterior
SD ratio and the requested \code{metaEntry}.
}
\description{
Increase in information density in prior vs posterior.
}
