% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MakeDGEList}
\alias{MakeDGEList}
\title{Create DGEList}
\usage{
MakeDGEList(dataset, samples, forwardSamples)
}
\arguments{
\item{dataset}{matrix with count data. Samples on rows and variants over columns}

\item{samples}{vector of sample names}

\item{forwardSamples}{vector of forward read files}
}
\value{
DGEList
}
\description{
Create DGEList
}
