% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeSlurm.R
\name{MakeSlurm}
\alias{MakeSlurm}
\title{Create a slurm job
Constructs and submits a slurm job to complete an analysis}
\usage{
MakeSlurm(
  pID,
  scriptID,
  ml = FALSE,
  ess = EssTarget(pID, scriptID),
  replace = FALSE,
  generous = TRUE
)
}
\arguments{
\item{pID}{Character specifying the project identifier to analyse}

\item{scriptID}{Character specifying the script to run}

\item{ml}{Logical; if true, calculate the marginal likelihood; if false,
run MCMCMC analysis}

\item{ess}{Numeric: Estimated sample size to attempt to obtain}

\item{replace}{Logical: should a corresponding job be cancelled in order to
run this job instead?}
}
\description{
Create a slurm job
Constructs and submits a slurm job to complete an analysis
}
