% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilePaths.R
\name{MatrixFile}
\alias{MatrixFile}
\title{Path to matrix file}
\usage{
MatrixFile(pID = "", ext = "nex")
}
\arguments{
\item{pID}{Character specifying the project identifier to analyse}

\item{ext}{File extension (default \code{"nex"}).}
}
\value{
Character string giving path to matrix file.
}
\description{
Returns the location of a project's morphological matrix file.
}
\seealso{
Other file-path helpers: 
\code{\link{AnalysisDir}()},
\code{\link{ConvergenceFile}()},
\code{\link{DiskFile}()},
\code{\link{DistanceFile}()},
\code{\link{DummyFile}()},
\code{\link{MarginalFile}()},
\code{\link{MemFile}()},
\code{\link{PFiles}()},
\code{\link{PPSamplerFile}()},
\code{\link{ParameterFile}()},
\code{\link{ParsEvalFile}()},
\code{\link{ParsimonyFile}()},
\code{\link{ResultsDir}()},
\code{\link{ScriptBase}()},
\code{\link{ScriptFile}()},
\code{\link{SlurmFile}()},
\code{\link{SlurmTemplate}()},
\code{\link{StoneFile}()},
\code{\link{StoneOrigin}()},
\code{\link{TimeFile}()},
\code{\link{TreeFiles}()},
\code{\link{TreeSampleFile}()}
}
\concept{file-path helpers}
