% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blasttools.R
\name{MultiBlaster}
\alias{MultiBlaster}
\title{Facilitate blast analysis of all sequences in fasta-file}
\usage{
MultiBlaster(fastaFile, seqNumber = 0, resultNumber = 5, viewChoice = "")
}
\arguments{
\item{fastaFile}{file with fasta sequences}

\item{seqNumber}{number of sequences from fasta to analys if set to
0 all sequences in the file are extracted.}

\item{resultNumber}{the number of blast hits per sequence to retain}

\item{viewChoice}{if set to yes/Yes/YES a view of the dataframe with blast
results will be shown using the function View.}
}
\value{
blast results for sequences in a dataframe
}
\description{
Parsing every sequence entry in a fasta file and use the
OnlineBlaster function to blast individual sequences directly on
NCBI. This function works with any number of sequences, but is
perhaps most suitable with less than 150 sequences. Consider using
local blast database for larger datasets.
}
\examples{
Fastafile <- system.file("extdata", "test.fa", package = "MetaBAnalysis")
MultiBlaster(fastaFile = Fastafile)

}
