% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blasttools.R
\name{OnlineBlaster}
\alias{OnlineBlaster}
\title{Blast a given nucleotide sequence at NCBI}
\usage{
OnlineBlaster(nucleotide)
}
\arguments{
\item{nucleotide}{string of DNA sequence that should be submitted to
blast}
}
\value{
dataframe of blast hits for @param nucleotide. If
the sequence is not similar to any known sequence at NCBI this will be
reported and an dataframe with NA values will be returned.
}
\description{
This facilitate the submission of a nucleotide sequence to NCBIs
blast database. Will fetch blast hits and return these hits as
a dataframe. This function largely replaces the need to have a
local blast database to compare to. If you have an active internet
access this method is strongly preferred to any approach relying on
local databases.
}
\details{
NB! The function needs internet access to work.
}
\examples{
ntExample <- sample(LETTERS[c(1, 3, 7, 20)], size = 100, replace = TRUE)
ntExample <- paste(ntExample, sep = "", collapse = "")
OnlineBlaster(ntExample)
}
