% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{OutCombine}
\alias{OutCombine}
\title{Summarize rows of forward and reverse counts per sample}
\usage{
OutCombine(dataset, samples)
}
\arguments{
\item{dataset}{a matrix with number of reads generated by the filterAndTrim function
from dada2}

\item{samples}{vector with names of samples to be merged}
}
\value{
matrix with number of reads before and after trimming per sample
}
\description{
Summarize rows of forward and reverse counts per sample
}
