% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{Panel}
\alias{Panel}
\title{Add label for plot region}
\usage{
Panel(i, xOffset = 3, yOffset = 0)
}
\arguments{
\item{i}{Index of label (1 = "(a)", or string to print}
}
\description{
Add label for plot region
}
