% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilePaths.R
\name{ParsEvalFile}
\alias{ParsEvalFile}
\title{Path to parsimony-evaluation results}
\usage{
ParsEvalFile(pID, scriptID = NULL, nPars, nBayes)
}
\arguments{
\item{pID}{Character specifying the project identifier to analyse}

\item{scriptID}{Character specifying the script to run}

\item{nPars, nBayes}{Integer counts of parsimony and Bayesian replicates.}
}
\value{
Character string giving \code{.rda} file path.
}
\description{
Constructs the path to the \code{.rda} file storing results of parsimony and
Bayesian evaluation for a given project, script, and parameter combination.
}
\seealso{
Other file-path helpers: 
\code{\link{AnalysisDir}()},
\code{\link{ConvergenceFile}()},
\code{\link{DiskFile}()},
\code{\link{DistanceFile}()},
\code{\link{DummyFile}()},
\code{\link{MarginalFile}()},
\code{\link{MatrixFile}()},
\code{\link{MemFile}()},
\code{\link{PFiles}()},
\code{\link{PPSamplerFile}()},
\code{\link{ParameterFile}()},
\code{\link{ParsimonyFile}()},
\code{\link{ResultsDir}()},
\code{\link{ScriptBase}()},
\code{\link{ScriptFile}()},
\code{\link{SlurmFile}()},
\code{\link{SlurmTemplate}()},
\code{\link{StoneFile}()},
\code{\link{StoneOrigin}()},
\code{\link{TimeFile}()},
\code{\link{TreeFiles}()},
\code{\link{TreeSampleFile}()}
}
\concept{file-path helpers}
