% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-experiments.R
\name{plot_summary}
\alias{plot_summary}
\title{Plots summary level information as a time series}
\usage{
plot_summary(
  feature,
  ...,
  mapping = .gg_check_for_aes(...),
  facet = TRUE,
  policy = TRUE,
  spaghetti = TRUE,
  points = FALSE
)
}
\arguments{
\item{feature}{The summary feature to plot. This is a column name of the \code{tables$summary}
data frame, it can include things like \code{incidence}, \code{cumulativeDeaths}}

\item{...}{ggplot format options which can be used to control the plot and
faceting}

\item{mapping}{ggplot \code{aes} mappings which can be used to override the defaults}

\item{facet}{automatically facet the plot base on the experimental design
if you are not happy with the defaults set this to \code{FALSE} and add your own
\code{facet_wrap()} command}

\item{policy}{Should we include a background layer showing the policy state
at any given point in time (see \code{plot_policy()})?}

\item{spaghetti}{Should we display multiple time-series as a spaghetti plot
(each simulation plotted individually as a line) or as a ribbon (mean
value and 0.95 quantiles for all replicas of a given model and experiment)}

\item{points}{Should we plot the data points as well as the ribbons / lines.}
}
\value{
a ggplot
}
\description{
Plots summary level information as a time series
}
\examples{
set_results_dir(here("../scratch/examples/lockdown-compliance"))
get_table("summary") \%>\% dplyr::glimpse()
tables \%>\% plot_summary(averageMobility, spaghetti=FALSE)
tables \%>\% plot_summary(averageViralLoad, spaghetti=FALSE)
tables \%>\% plot_summary(cumulativeDeaths, spaghetti=FALSE)
tables \%>\% plot_summary(hospitalisedCount, spaghetti=FALSE)
tables \%>\% plot_summary(cumulativeInfections, spaghetti=FALSE)
tables \%>\% plot_summary(cumulativeMobilityDecrease, spaghetti=FALSE)
}
