% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parsimony.R
\name{PosteriorTreeSteps}
\alias{PosteriorTreeSteps}
\title{Parsimony lenths of trees in posterior distribution}
\usage{
PosteriorTreeSteps(projects, models)
}
\description{
Parsimony lenths of trees in posterior distribution
}
