% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RemoveLowFreqSeqs}
\alias{RemoveLowFreqSeqs}
\title{Remove species with counts below a given threshold}
\usage{
RemoveLowFreqSeqs(dataset, threshold, subValue = 0)
}
\arguments{
\item{dataset}{a dataframe where the first column contains
species names and the rest of the dataframe contains integers >= 0}

\item{threshold}{cut-off value in percent}

\item{subValue}{value to replace data below threshold with}
}
\value{
dataframe with filtered data removed.
}
\description{
Filter data below a certain threshold
}
\examples{
testdata <- data.frame(Species = c("Esox lucius",
                                   "Perca fluviatilis",
                                   "Tinca tinca"),
                       Random1 = LETTERS[1:3],
                       Random2 = LETTERS[1:3],
                       Random3 = LETTERS[1:3],
                       Sample1 = c(10010, 3921, 2),
                       Sample2 = c(9900, 5, 1301))
RemoveLowFreqSeqs(dataset = testdata,
                   threshold = 0.1,
                   subValue = 0)

}
