% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilePaths.R
\name{ScriptFile}
\alias{ScriptFile}
\title{Path to RevBayes script}
\usage{
ScriptFile(pID, scriptID, analysis = "mcmcmc")
}
\arguments{
\item{pID}{Character specifying the project identifier to analyse}

\item{scriptID}{Character specifying the script to run}

\item{analysis}{One of \code{"mcmcmc"}, \code{"marginal"}, or \code{"ppsample"}.}
}
\value{
Character string giving file path.
}
\description{
Constructs the path to a \code{.Rev} analysis script within the corresponding
RevBayes repository.
}
\seealso{
Other file-path helpers: 
\code{\link{AnalysisDir}()},
\code{\link{ConvergenceFile}()},
\code{\link{DiskFile}()},
\code{\link{DistanceFile}()},
\code{\link{DummyFile}()},
\code{\link{MarginalFile}()},
\code{\link{MatrixFile}()},
\code{\link{MemFile}()},
\code{\link{PFiles}()},
\code{\link{PPSamplerFile}()},
\code{\link{ParameterFile}()},
\code{\link{ParsEvalFile}()},
\code{\link{ParsimonyFile}()},
\code{\link{ResultsDir}()},
\code{\link{ScriptBase}()},
\code{\link{SlurmFile}()},
\code{\link{SlurmTemplate}()},
\code{\link{StoneFile}()},
\code{\link{StoneOrigin}()},
\code{\link{TimeFile}()},
\code{\link{TreeFiles}()},
\code{\link{TreeSampleFile}()}
}
\concept{file-path helpers}
