% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilePaths.R
\name{SlurmFile}
\alias{SlurmFile}
\title{Path to generated SLURM script}
\usage{
SlurmFile(pID, scriptID, ml = FALSE)
}
\arguments{
\item{pID}{Character specifying the project identifier to analyse}

\item{scriptID}{Character specifying the script to run}

\item{ml}{Logical; whether this corresponds to a marginal-likelihood run.}
}
\value{
Character string giving file path.
}
\description{
Constructs the path to the SLURM job script for a given project and script.
}
\seealso{
Other file-path helpers: 
\code{\link{AnalysisDir}()},
\code{\link{ConvergenceFile}()},
\code{\link{DiskFile}()},
\code{\link{DistanceFile}()},
\code{\link{DummyFile}()},
\code{\link{MarginalFile}()},
\code{\link{MatrixFile}()},
\code{\link{MemFile}()},
\code{\link{PFiles}()},
\code{\link{PPSamplerFile}()},
\code{\link{ParameterFile}()},
\code{\link{ParsEvalFile}()},
\code{\link{ParsimonyFile}()},
\code{\link{ResultsDir}()},
\code{\link{ScriptBase}()},
\code{\link{ScriptFile}()},
\code{\link{SlurmTemplate}()},
\code{\link{StoneFile}()},
\code{\link{StoneOrigin}()},
\code{\link{TimeFile}()},
\code{\link{TreeFiles}()},
\code{\link{TreeSampleFile}()}
}
\concept{file-path helpers}
