% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilePaths.R
\name{SlurmTemplate}
\alias{SlurmTemplate}
\title{Path to SLURM submission template}
\usage{
SlurmTemplate(ml, mpi = TRUE)
}
\arguments{
\item{ml}{Logical; whether this is a marginal-likelihood run.}

\item{mpi}{Logical; whether the analysis uses MPI parallelization.}
}
\value{
Character string giving path to SLURM template.
}
\description{
Returns the template script appropriate for a given analysis mode.
}
\seealso{
Other file-path helpers: 
\code{\link{AnalysisDir}()},
\code{\link{ConvergenceFile}()},
\code{\link{DiskFile}()},
\code{\link{DistanceFile}()},
\code{\link{DummyFile}()},
\code{\link{MarginalFile}()},
\code{\link{MatrixFile}()},
\code{\link{MemFile}()},
\code{\link{PFiles}()},
\code{\link{PPSamplerFile}()},
\code{\link{ParameterFile}()},
\code{\link{ParsEvalFile}()},
\code{\link{ParsimonyFile}()},
\code{\link{ResultsDir}()},
\code{\link{ScriptBase}()},
\code{\link{ScriptFile}()},
\code{\link{SlurmFile}()},
\code{\link{StoneFile}()},
\code{\link{StoneOrigin}()},
\code{\link{TimeFile}()},
\code{\link{TreeFiles}()},
\code{\link{TreeSampleFile}()}
}
\concept{file-path helpers}
