% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SpeciesPercent}
\alias{SpeciesPercent}
\title{SpeciesPercent}
\usage{
SpeciesPercent(dataFrame)
}
\arguments{
\item{dataFrame}{a dataframe containing counts of species with the species names in the first column to the left.}
}
\value{
a vector of species percentages.
}
\description{
Takes a dataframe of species counts and returns a vector with the percentage of counts per species.
}
