% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{toRad}
\alias{toRad}
\title{This function takes in one parameter:degrees.
It then calculates the radian of that specified value.}
\usage{
toRad(degrees)
}
\arguments{
\item{degrees}{that you want to convert to radian}
}
\value{
it returns the radian value of that degree value
}
\description{
This function takes in one parameter:degrees.
It then calculates the radian of that specified value.
}
