\name{toRad}
\alias{toRad}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Degrees to Radian}
\description{
This function takes in one parameter:degrees.
It then calculates the radian of that specified value.
}
\usage{toRad(x)}
\arguments{
  \item{x}{a degree value}
}
\value{the radian value of the degrees value}
