% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilePaths.R
\name{TreeFiles}
\alias{TreeFiles}
\title{List tree-file paths}
\usage{
TreeFiles(pID, scriptID, compressed = FALSE)
}
\arguments{
\item{pID}{Character specifying the project identifier to analyse}

\item{scriptID}{Character specifying the script to run}

\item{compressed}{Logical; whether to list compressed files.}
}
\value{
Character vector of tree-file paths.
}
\description{
Returns a vector of paths to \code{.trees} (or compressed \code{.tar.gz}) files
generated by an analysis.
}
\seealso{
Other file-path helpers: 
\code{\link{AnalysisDir}()},
\code{\link{ConvergenceFile}()},
\code{\link{DiskFile}()},
\code{\link{DistanceFile}()},
\code{\link{DummyFile}()},
\code{\link{MarginalFile}()},
\code{\link{MatrixFile}()},
\code{\link{MemFile}()},
\code{\link{PFiles}()},
\code{\link{PPSamplerFile}()},
\code{\link{ParameterFile}()},
\code{\link{ParsEvalFile}()},
\code{\link{ParsimonyFile}()},
\code{\link{ResultsDir}()},
\code{\link{ScriptBase}()},
\code{\link{ScriptFile}()},
\code{\link{SlurmFile}()},
\code{\link{SlurmTemplate}()},
\code{\link{StoneFile}()},
\code{\link{StoneOrigin}()},
\code{\link{TimeFile}()},
\code{\link{TreeSampleFile}()}
}
\concept{file-path helpers}
