% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{WriteSeqsPerSample}
\alias{WriteSeqsPerSample}
\title{Write Sequences Belonging To Each Sample To Individual Fasta Files}
\usage{
WriteSeqsPerSample(countsObject = yAll$counts, fileSuffix = "")
}
\arguments{
\item{countsObject}{object with columns of samples with counts and rows of sequences (where each row name is a specific sequence). This corresponds to yAll$counts generated by MakeDGEList()}

\item{fileSuffix}{string of text that will be placed at the end of the generated filenames: \link[=fileSuffix]{sample name}.fa}
}
\description{
Write Sequences Belonging To Each Sample To Individual Fasta Files
}
